refactor(dashboard): extract attack card renderers into per-attack mixin modules
