+++ hackagent/server/dashboard/_page.py
@@ -8,7 +8,6 @@ from __future__ import annotations
 import asyncio
 from collections import defaultdict
 import contextlib
-import html
 import json
 import math
 import re
@@ -34,6 +33,18 @@ from ._helpers import (
     _result_bucket,
     _short_date,
 )
+from .attack_cards import (
+    AttackCardSharedMixin,
+    BaselineCardMixin,
+    BonCardMixin,
+    PairCardMixin,
+    AutodanCardMixin,
+    AdvprefixCardMixin,
+    PapCardMixin,
+    TapCardMixin,
+    GenericCardMixin,
+    MmlCardMixin,
+)
 
 _VIEW_LABELS = {
     "dashboard": "Dashboard",
@@ -46,7 +57,18 @@ _DASHBOARD_RUN_SCAN_LIMIT = 10
 _RUNS_VIEW_PAGE_SIZE = 15
 
 
-class DashboardPage:
+class DashboardPage(
+    AttackCardSharedMixin,
+    BaselineCardMixin,
+    BonCardMixin,
+    PairCardMixin,
+    AutodanCardMixin,
+    AdvprefixCardMixin,
+    PapCardMixin,
+    TapCardMixin,
+    GenericCardMixin,
+    MmlCardMixin,
+):
     """Owns all NiceGUI widgets for a single dashboard page request.
 
     A new instance is created inside the ``@ui.page("/")`` handler for every
@@ -1102,6 +1124,8 @@ function hackAgentCopyFallback(text) {
             self._render_advprefix_goal_card(row, _pr, _gs, detail_mode=True)
         elif ha == "autodanturbo":
             self._render_autodan_goal_card(row, data, detail_mode=True)  # type: ignore[arg-type]
+        elif ha == "mml":
+            self._render_mml_goal_card(row, data, detail_mode=True)  # type: ignore[arg-type]
         else:
             _req, _resp, _gr_evt = data  # type: ignore[misc]
             self._render_generic_goal_card(
@@ -1348,6 +1372,12 @@ function hackAgentCopyFallback(text) {
                 "text-[10px] font-semibold tracking-widest text-grey-5 uppercase"
             )
             ui.label(notes).classes("text-sm")
+
+        # MML-specific rendering: Image + Prompt + Response
+        metadata = row.get("metadata") if isinstance(row.get("metadata"), dict) else {}
+        encoding_mode = metadata.get("encoding_mode")
+        if encoding_mode:
+            self._render_mml_result_section(row, metadata)
 
         # Key-value detail table
         detail_fields = self._build_result_detail_fields(row)
@@ -3726,65 +3756,8 @@ function hackAgentCopyFallback(text) {
             "created_at": result.get("updated_at") or result.get("created_at"),
         }
 
-    @staticmethod
-    def _extract_request_response_candidates(content: object) -> tuple[object, object]:
-        """Best-effort extraction of request/response payloads from trace content."""
-        if not isinstance(content, dict):
-            return None, None
-
-        metadata = (
-            content.get("metadata") if isinstance(content.get("metadata"), dict) else {}
-        )
-        nested_result = (
-            content.get("result") if isinstance(content.get("result"), dict) else {}
-        )
-
-        request_value = (
-            content.get("request")
-            or content.get("prefix")
-            or content.get("prompt")
-            or nested_result.get("request")
-            or nested_result.get("prefix")
-            or nested_result.get("prompt")
-            or metadata.get("request")
-            or metadata.get("prefix")
-            or metadata.get("prompt")
-        )
-        response_value = (
-            content.get("response")
-            or content.get("completion")
-            or content.get("answer")
-            or nested_result.get("response")
-            or nested_result.get("completion")
-            or nested_result.get("answer")
-            or metadata.get("response")
-            or metadata.get("completion")
-            or metadata.get("answer")
-            or metadata.get("raw_response_body")
-        )
-
-        if isinstance(request_value, dict):
-            request_value = (
-                request_value.get("prompt")
-                or request_value.get("request")
-                or request_value
-            )
-
-        if isinstance(response_value, dict):
-            response_value = (
-                response_value.get("target_response")
-                or response_value.get("response")
-                or response_value.get("completion")
-                or response_value.get("generated_text")
-                or response_value
-            )
-
-        return request_value, response_value
-
     def _ensure_evaluation_request_response(
-        self,
-        serialized_traces: list[dict],
-        result: dict,
+        self, serialized_traces: list[dict], result: dict
     ) -> list[dict]:
         """Inject Request/Response in evaluation traces so they are always visible."""
 
@@ -3949,2566 +3922,6 @@ function hackAgentCopyFallback(text) {
 
         return serialized_traces
 
-    # ── Attack-specific goal card helpers ─────────────────────────────────────
-
-    @staticmethod
-    def _border_color_for_bucket(bucket: str) -> str:
-        if bucket == "jailbreak":
-            return "border-red-400"
-        if bucket == "mitigated":
-            return "border-green-400"
-        if bucket == "failed":
-            return "border-orange-400"
-        return "border-grey-300"
-
-    def _render_compact_card(self, row: dict, on_click) -> None:
-        """Render a compact clickable goal card for the left-panel list view."""
-        goal_text = str(row.get("goal") or "—")
-        goal_number = row.get("goal_number", "?")
-        bucket = row.get("_bucket", "pending")
-        border_color = self._border_color_for_bucket(bucket)
-        with (
-            ui.card()
-            .tight()
-            .classes(
-                f"w-full border-l-4 {border_color} cursor-pointer"
-                " hover:shadow-sm transition-shadow"
-            )
-            .on("click", on_click)
-        ):
-            with ui.row().classes("items-start gap-2 px-3 py-2 w-full"):
-                ui.label(f"#{goal_number}").classes(
-                    "font-bold text-xs text-grey-5 shrink-0 w-6 pt-0.5 text-right"
-                )
-                ui.label(goal_text).classes(
-                    "text-xs text-grey-8 flex-1 leading-snug whitespace-pre-wrap"
-                )
-
-    @contextlib.contextmanager
-    def _goal_card_shell(self, row: dict, detail_mode: bool = False):
-        goal_text = str(row.get("goal") or "—")
-        goal_number = row.get("goal_number", "?")
-        bucket = row.get("_bucket", "pending")
-        border_color = self._border_color_for_bucket(bucket)
-        if detail_mode:
-            _cat = row.get("_goal_category") or ""
-            _subcat = row.get("_goal_subcategory") or ""
-            with ui.column().classes("w-full gap-2"):
-                with ui.row().classes("items-center gap-2 flex-wrap"):
-                    ui.label(f"Goal #{goal_number}").classes(
-                        "font-bold text-base shrink-0"
-                    )
-                    if bucket == "jailbreak":
-                        ui.badge("Jailbreak", color="negative").classes("text-xs")
-                    elif bucket == "mitigated":
-                        ui.badge("Mitigated", color="positive").classes("text-xs")
-                    elif bucket == "failed":
-                        ui.badge("Error", color="warning").classes("text-xs")
-                    lat = row.get("_goal_latency")
-                    if lat and lat != "—":
-                        ui.badge(f"Latency: {lat}", color="grey-7").classes("text-xs")
-                if _cat:
-                    _cat_str = _cat
-                    if _subcat and _subcat not in ("", "N/A"):
-                        _cat_str += f" › {_subcat}"
-                    ui.label(_cat_str).classes("text-xs text-grey-5 tracking-wide")
-                ui.label(goal_text).classes(
-                    "text-sm text-grey-8 whitespace-pre-wrap leading-relaxed"
-                )
-            ui.separator().classes("my-2")
-            yield
-        else:
-            with ui.card().tight().classes(f"w-full border-l-4 {border_color}"):
-                with ui.column().classes("w-full gap-2 p-3") as col:
-                    ui.label(f"Goal #{goal_number}").classes(
-                        "font-bold text-sm shrink-0"
-                    )
-                    ui.label(goal_text).classes(
-                        "text-sm text-grey-8 whitespace-pre-wrap"
-                    )
-                    yield col
-
-    @staticmethod
-    def _wire_expand_toggle(body_col) -> None:
-        toggle_btn = (
-            ui.button("Expand", icon="expand_more")
-            .props("flat no-caps size=sm color=grey-7")
-            .classes("w-full")
-        )
-        _state: dict = {"open": False}
-
-        def _toggle(_b=body_col, _btn=toggle_btn, _s=_state) -> None:
-            _s["open"] = not _s["open"]
-            _b.set_visibility(_s["open"])
-            _btn.props(
-                f"label={'Collapse' if _s['open'] else 'Expand'} icon={'expand_less' if _s['open'] else 'expand_more'} flat no-caps size=sm color=grey-7"
-            )
-
-        toggle_btn.on_click(_toggle)
-
-    # ── Baseline ──────────────────────────────────────────────────────────────
-
-    @staticmethod
-    def _parse_baseline_traces(traces: list[dict], goal: str = "") -> list[dict]:
-        """Parse Baseline attack traces into per-template rows.
-
-        Each row dict:
-          num           – 1-based row number
-          template      – attack_prompt with goal text replaced by {goal}
-          prompt        – raw attack prompt sent to target
-          response      – target model response
-          result        – "Jailbreak" | "Mitigated" | "Error"
-          _bucket       – "jailbreak" | "mitigated" | "error"
-        """
-        from collections import deque  # noqa: PLC0415
-
-        sorted_traces = sorted(traces, key=lambda x: x.get("sequence", 0))
-
-        # Baseline writes one "Template:…" interaction trace per template and
-        # one evaluation trace from the baseline_pattern_evaluator.
-        interaction_traces: list[tuple[str, dict]] = []
-        eval_trace_result: dict = {}
-
-        for td in sorted_traces:
-            content = td.get("content") or {}
-            step_name = str(content.get("step_name") or td.get("step_name") or "")
-            if step_name.startswith("Template:"):
-                interaction_traces.append((step_name, content))
-            elif str(content.get("evaluator") or "").startswith("baseline_"):
-                eval_trace_result = content.get("result") or {}
-
-        # Stable sort: by step_name then first message content
-        interaction_traces.sort(
-            key=lambda x: (
-                x[0],
-                (x[1].get("request") or {}).get("messages", [{}])[0].get("content", "")
-                if (x[1].get("request") or {}).get("messages")
-                else "",
-            )
-        )
-
-        # Build lookup: (template_category, response_length) → deque of eval entries
-        eval_by_key: dict[tuple, deque] = {}
-        for ev in eval_trace_result.get("evaluations") or []:
-            key = (
-                ev.get("template_category") or "",
-                int(ev.get("response_length") or 0),
-            )
-            if key not in eval_by_key:
-                eval_by_key[key] = deque()
-            eval_by_key[key].append(ev)
-
-        rows: list[dict] = []
-        for idx, (_, content) in enumerate(interaction_traces, start=1):
-            request = content.get("request") or {}
-            messages = request.get("messages") or []
-            attack_prompt = messages[0].get("content", "") if messages else ""
-            if not attack_prompt:
-                attack_prompt = str(request.get("prompt") or "")
-
-            _raw_resp = content.get("response")
-            _actual_resp, _g_side, _g_expl, _g_cats = (
-                DashboardPage._extract_guardrail_from_response(_raw_resp)
-            )
-            response_text = str(_actual_resp or "")
-
-            metadata = content.get("metadata") or {}
-            template_category = str(metadata.get("template_category") or "")
-            response_length = int(metadata.get("response_length") or len(response_text))
-
-            # Replace goal text in prompt with {goal} placeholder for display
-            if goal and goal in attack_prompt:
-                template_display = attack_prompt.replace(goal, "{goal}", 1)
-            else:
-                template_display = attack_prompt
-
-            # Match against the evaluator result
-            key = (template_category, response_length)
-            success: bool | None = None
-            q = eval_by_key.get(key)
-            if q:
-                ev = q.popleft()
-                success = bool(ev.get("success", False))
-
-            if _g_side:
-                bucket = "mitigated"
-            elif success is True:
-                bucket = "jailbreak"
-            elif success is False:
-                bucket = "mitigated"
-            elif response_text:
-                # Have a response but no evaluator match — treat as pending
-                bucket = "mitigated"
-            else:
-                bucket = "error"
-
-            rows.append(
-                {
-                    "num": idx,
-                    "template": template_display,
-                    "template_category": template_category,
-                    "prompt": attack_prompt,
-                    "response": response_text,
-                    "result": (
-                        "Jailbreak"
-                        if bucket == "jailbreak"
-                        else "Error"
-                        if bucket == "error"
-                        else "Mitigated"
-                    ),
-                    "_bucket": bucket,
-                    "_guardrail_side": _g_side,
-                    "_guardrail_explanation": _g_expl,
-                    "_guardrail_categories": _g_cats,
-                }
-            )
-
-        return rows
-
-    def _render_baseline_goal_card(
-        self, row: dict, template_rows: list[dict], detail_mode: bool = False
-    ) -> None:
-        """Render a Baseline goal card grouped by template category."""
-
-        def _fmt_cat(cat: str) -> str:
-            return cat.replace("_", " ").title() if cat else "Uncategorised"
-
-        # Build category groups preserving insertion order
-        groups: dict[str, list[dict]] = {}
-        for tr in template_rows:
-            cat = tr.get("template_category") or ""
-            groups.setdefault(cat, []).append(tr)
-
-        n_jailbreaks = sum(1 for r in template_rows if r["_bucket"] == "jailbreak")
-        n_mitigated = sum(1 for r in template_rows if r["_bucket"] == "mitigated")
-        n_errors = sum(1 for r in template_rows if r["_bucket"] == "error")
-
-        bl_cols = [
-            {
-                "name": "template_short",
-                "label": "Template",
-                "field": "template_short",
-                "align": "left",
-            },
-            {
-                "name": "result",
-                "label": "Result",
-                "field": "result",
-                "align": "center",
-                "style": "width:100px",
-            },
-        ]
-
-        with self._goal_card_shell(row, detail_mode):
-            if not template_rows:
-                ui.label("No Baseline template data recorded.").classes(
-                    "text-sm text-grey-6"
-                )
-                return
-
-            with ui.column().classes("w-full gap-2 mt-1") as body_col:
-                if not detail_mode:
-                    body_col.set_visibility(False)
-
-                for cat, rows_in_cat in groups.items():
-                    cat_label = _fmt_cat(cat)
-                    cat_n_jailbreaks = sum(
-                        1 for r in rows_in_cat if r["_bucket"] == "jailbreak"
-                    )
-
-                    with ui.row().classes("items-center gap-2 mt-3 mb-0.5 px-1"):
-                        ui.label(cat_label).classes(
-                            "text-xs font-semibold text-grey-6 uppercase tracking-wide"
-                        )
-                        ui.badge(
-                            f"{len(rows_in_cat)} prompt{'s' if len(rows_in_cat) != 1 else ''}",
-                            color="grey-5",
-                        ).classes("text-xs")
-                        if cat_n_jailbreaks:
-                            ui.badge(
-                                f"{cat_n_jailbreaks} jailbreak{'s' if cat_n_jailbreaks != 1 else ''}",
-                                color="negative",
-                            ).classes("text-xs")
-
-                    tbl_rows = [
-                        {
-                            "_num": tr["num"],
-                            "template_short": (
-                                (
-                                    tr["template"].replace("{goal}", "", 1)[:80]
-                                    + "\u2026"
-                                )
-                                if len(tr["template"].replace("{goal}", "", 1)) > 80
-                                else tr["template"].replace("{goal}", "", 1)
-                            )
-                            or "—",
-                            "result": tr["result"],
-                            "_bucket": tr["_bucket"],
-                            "_full_prompt": tr.get("prompt") or "",
-                            "_response": tr.get("response") or "",
-                            "_guardrail_side": tr.get("_guardrail_side") or "",
-                            "_guardrail_explanation": tr.get("_guardrail_explanation")
-                            or "",
-                            "_guardrail_categories": tr.get("_guardrail_categories")
-                            or [],
-                        }
-                        for tr in rows_in_cat
-                    ]
-
-                    tbl = (
-                        ui.table(
-                            columns=bl_cols,
-                            rows=tbl_rows,
-                            row_key="_num",
-                        )
-                        .classes("w-full text-xs")
-                        .props("dense flat")
-                    )
-                    if detail_mode:
-                        tbl.props(
-                            f":expanded-rows='{json.dumps([r['_num'] for r in tbl_rows])}'"
-                        )
-
-                    tbl.add_slot(
-                        "body",
-                        r"""
-<q-tr :props="props" @click="props.expand = !props.expand" style="cursor:pointer">
-  <q-td key="template_short" :props="props"
-        style="white-space:pre-wrap;word-break:break-word;max-width:360px">
-    {{ props.row.template_short }}
-  </q-td>
-  <q-td key="result" :props="props">
-    <q-badge v-if="props.row._bucket === 'jailbreak'" color="negative" class="text-xs">Jailbreak</q-badge>
-    <q-badge v-else-if="props.row._bucket === 'mitigated'" color="positive" class="text-xs">Mitigated</q-badge>
-    <q-badge v-else color="warning" class="text-xs">Error</q-badge>
-  </q-td>
-</q-tr>
-<q-tr v-show="props.expand" :props="props" @click="props.expand = false" style="cursor:pointer">
-  <q-td colspan="100%" class="bg-grey-1">
-    <div class="q-pa-sm">
-      <div class="row items-center justify-between q-mb-xs">
-        <span class="text-caption text-weight-bold text-uppercase text-grey-6">PROMPT SENT TO TARGET</span>
-        <span :data-cliptext="props.row._full_prompt||''" onclick="event.stopPropagation();var s=this,t=s.dataset.cliptext||'',ic=s.querySelector('.q-icon');if(window.navigator&&window.navigator.clipboard)window.navigator.clipboard.writeText(t);if(ic)ic.textContent='check';s.setAttribute('title','Copied to clipboard');setTimeout(function(){if(ic)ic.textContent='content_copy';s.setAttribute('title','Copy to clipboard');},2000);" style="display:inline-flex;cursor:pointer" title="Copy to clipboard"><q-btn flat dense size="xs" icon="content_copy" color="grey-6" /></span>
-      </div>
-      <pre style="font-size:11px;padding:8px;background:white;border:1px solid #e0e0e0;
-                  border-radius:4px;margin-bottom:8px;white-space:pre-wrap;word-break:break-word">{{ props.row._full_prompt || '\u2014' }}</pre>
-      <template v-if="props.row._guardrail_side !== 'before'">
-        <div class="row items-center justify-between q-mb-xs">
-          <span class="text-caption text-weight-bold text-uppercase text-grey-6">TARGET RESPONSE</span>
-          <span :data-cliptext="props.row._response||''" onclick="event.stopPropagation();var s=this,t=s.dataset.cliptext||'',ic=s.querySelector('.q-icon');if(window.navigator&&window.navigator.clipboard)window.navigator.clipboard.writeText(t);if(ic)ic.textContent='check';s.setAttribute('title','Copied to clipboard');setTimeout(function(){if(ic)ic.textContent='content_copy';s.setAttribute('title','Copy to clipboard');},2000);" style="display:inline-flex;cursor:pointer" title="Copy to clipboard"><q-btn flat dense size="xs" icon="content_copy" color="grey-6" /></span>
-        </div>
-        <pre style="font-size:11px;padding:8px;background:white;border:1px solid #e0e0e0;
-                    border-radius:4px;margin-bottom:8px;white-space:pre-wrap;word-break:break-word">{{ props.row._response || 'No response recorded.' }}</pre>
-      </template>
-      <div v-if="props.row._guardrail_side === 'before'" style="margin-bottom:8px">
-        <div class="text-caption text-weight-bold text-uppercase q-mb-xs" style="color:#c2410c">&#x26a0; BEFORE GUARDRAIL &#x2014; BLOCKED</div>
-        <pre style="font-size:11px;padding:10px;background:#fff7ed;border:2px solid #f97316;border-radius:4px;white-space:pre-wrap;word-break:break-word;margin:0"><span v-if="props.row._guardrail_categories && props.row._guardrail_categories.length" style="font-weight:700;color:#c2410c">Categories: </span><span v-if="props.row._guardrail_categories && props.row._guardrail_categories.length" style="color:#9a3412">{{ props.row._guardrail_categories.join(', ') }}</span><span v-if="props.row._guardrail_categories && props.row._guardrail_categories.length">&#10;&#10;</span><span style="font-weight:700;color:#c2410c">Explanation: </span><span style="color:#6b7280">{{ props.row._guardrail_explanation }}</span></pre>
-      </div>
-      <div v-else-if="props.row._guardrail_side === 'after'" style="margin-bottom:8px">
-        <div class="text-caption text-weight-bold text-uppercase q-mb-xs" style="color:#dc2626">&#x1f6ab; AFTER GUARDRAIL &#x2014; CENSORED</div>
-        <pre style="font-size:11px;padding:10px;background:#fef2f2;border:2px solid #ef4444;border-radius:4px;white-space:pre-wrap;word-break:break-word;margin:0"><span v-if="props.row._guardrail_categories && props.row._guardrail_categories.length" style="font-weight:700;color:#dc2626">Categories: </span><span v-if="props.row._guardrail_categories && props.row._guardrail_categories.length" style="color:#991b1b">{{ props.row._guardrail_categories.join(', ') }}</span><span v-if="props.row._guardrail_categories && props.row._guardrail_categories.length">&#10;&#10;</span><span style="font-weight:700;color:#dc2626">Explanation: </span><span style="color:#6b7280">{{ props.row._guardrail_explanation }}</span></pre>
-      </div>
-      <div v-else-if="props.row._guardrail_side" style="margin-bottom:8px">
-        <div class="text-caption text-weight-bold text-uppercase q-mb-xs" style="color:#616161">&#x1f6e1; GUARDRAIL &#x2014; BLOCKED</div>
-        <pre style="font-size:11px;padding:10px;background:#f5f5f5;border:2px solid #9e9e9e;border-radius:4px;white-space:pre-wrap;word-break:break-word;margin:0"><span v-if="props.row._guardrail_categories && props.row._guardrail_categories.length" style="font-weight:700;color:#616161">Categories: </span><span v-if="props.row._guardrail_categories && props.row._guardrail_categories.length" style="color:#374151">{{ props.row._guardrail_categories.join(', ') }}</span><span v-if="props.row._guardrail_categories && props.row._guardrail_categories.length">&#10;&#10;</span><span style="font-weight:700;color:#616161">Explanation: </span><span style="color:#6b7280">{{ props.row._guardrail_explanation }}</span></pre>
-      </div>
-    </div>
-  </q-td>
-</q-tr>
-""",
-                    )
-
-                # ── Summary row ───────────────────────────────────────
-                ui.separator().classes("mt-2")
-                with ui.row().classes("items-center gap-2 mt-2 flex-wrap"):
-                    ui.label("Summary:").classes("text-xs font-semibold text-grey-6")
-                    if n_jailbreaks:
-                        ui.badge(
-                            f"{n_jailbreaks} Jailbreak{'s' if n_jailbreaks != 1 else ''}",
-                            color="negative",
-                        ).classes("text-xs")
-                    ui.badge(f"{n_mitigated} Mitigated", color="positive").classes(
-                        "text-xs"
-                    )
-                    if n_errors:
-                        ui.badge(
-                            f"{n_errors} Error{'s' if n_errors != 1 else ''}",
-                            color="warning",
-                        ).classes("text-xs")
-
-            if not detail_mode:
-                self._wire_expand_toggle(body_col)
-
-    # ── Best-of-N (BoN) ───────────────────────────────────────────────────────
-
-    @staticmethod
-    def _parse_bon_traces(traces: list[dict]) -> list[dict]:
-        """Parse BoN traces into per-step groups.
-
-        Returns a list of step dicts, each containing:
-          step          – 0-based step index
-          step_label    – human label "Step N / M"
-          is_jailbreak  – True if the judge confirmed jailbreak for this step
-          candidates    – list of candidate dicts:
-              k                – 1-based candidate index
-              augmented_prompt – text sent to target
-              response         – target model response
-              response_length  – char length of response
-              is_best          – True if selected as best in the step
-              error            – error string or None
-              _guardrail_side  – "" | "before" | "after" | "unknown"
-              _guardrail_explanation – explanation string
-        """
-        candidate_traces: list[dict] = []
-        eval_traces: list[dict] = []
-
-        for td in sorted(traces, key=lambda x: x.get("sequence", 0)):
-            content = td.get("content") or {}
-            meta = content.get("metadata") or {}
-            dtype = str(meta.get("display_type") or "").lower()
-            if dtype == "bon_candidate":
-                candidate_traces.append(td)
-            elif dtype == "bon_evaluation":
-                eval_traces.append(td)
-
-        # Build lookup: step index → is_jailbreak (from evaluation traces only)
-        step_jailbreak: dict[int, bool] = {}
-        for td in eval_traces:
-            content = td.get("content") or {}
-            meta = content.get("metadata") or {}
-            s = meta.get("step")
-            if s is not None:
-                step_jailbreak[int(s)] = bool(meta.get("is_jailbreak", False))
-
-        # Group candidate traces by step
-        by_step: dict[int, list[dict]] = {}
-        for td in candidate_traces:
-            content = td.get("content") or {}
-            meta = content.get("metadata") or {}
-            s = int(meta.get("step", 0))
-            by_step.setdefault(s, []).append(td)
-
-        if not by_step:
-            return []
-
-        n_steps_seen = 1
-        for td in candidate_traces:
-            content = td.get("content") or {}
-            meta = content.get("metadata") or {}
-            n_steps_seen = max(n_steps_seen, int(meta.get("n_steps", 1)))
-
-        steps: list[dict] = []
-        for s in sorted(by_step.keys()):
-            cands = []
-            for td in sorted(
-                by_step[s],
-                key=lambda x: int(
-                    (x.get("content") or {})
-                    .get("metadata", {})
-                    .get("candidate_index", 0)
-                ),
-            ):
-                content = td.get("content") or {}
-                meta = content.get("metadata") or {}
-                request = content.get("request") or {}
-                response_obj = content.get("response")
-
-                augmented_prompt = (
-                    request.get("prompt")
-                    or (request.get("messages") or [{}])[0].get("content", "")
-                    if isinstance(request, dict)
-                    else ""
-                )
-
-                response_obj, _g_side, _g_expl, _g_cats = (
-                    DashboardPage._extract_guardrail_from_response(response_obj)
-                )
-
-                if isinstance(response_obj, dict):
-                    response_text = (
-                        response_obj.get("generated_text")
-                        or response_obj.get("completion")
-                        or ""
-                    )
-                    error_text = response_obj.get("error_message")
-                elif response_obj is not None:
-                    response_text = str(response_obj)
-                    error_text = None
-                else:
-                    response_text = ""
-                    error_text = None
-
-                resp_len = int(meta.get("response_length", len(response_text or "")))
-                cands.append(
-                    {
-                        "k": int(meta.get("candidate_index", 0)),
-                        "augmented_prompt": augmented_prompt,
-                        "response": response_text,
-                        "response_length": resp_len,
-                        "is_best": bool(meta.get("is_best", False)),
-                        "error": error_text,
-                        "_guardrail_side": _g_side,
-                        "_guardrail_explanation": _g_expl,
-                        "_guardrail_categories": _g_cats,
-                    }
-                )
-
-            steps.append(
-                {
-                    "step": s,
-                    "step_label": f"Step {s + 1} / {n_steps_seen}",
-                    "is_jailbreak": step_jailbreak.get(s, False),
-                    "candidates": cands,
-                }
-            )
-
-        return steps
-
-    def _render_bon_goal_card(
-        self, row: dict, step_groups: list[dict], detail_mode: bool = False
-    ) -> None:
-        """Render a BoN goal card with per-step candidate tables."""
-        with self._goal_card_shell(row, detail_mode):
-            if not step_groups:
-                ui.label("No BoN step results recorded.").classes("text-sm text-grey-6")
-                return
-
-            with ui.column().classes("w-full gap-2 mt-1") as body_col:
-                if not detail_mode:
-                    body_col.set_visibility(False)
-
-                for sg in step_groups:
-                    step_label = sg["step_label"]
-                    is_jailbreak_step = sg["is_jailbreak"]
-                    candidates = sg["candidates"]
-
-                    with ui.row().classes("items-center gap-2 mt-3 mb-0.5 px-1"):
-                        ui.label(step_label).classes(
-                            "text-xs font-semibold text-grey-6 uppercase tracking-wide"
-                        )
-                        if is_jailbreak_step:
-                            ui.badge("Jailbreak", color="negative").classes("text-xs")
-
-                    columns = [
-                        {
-                            "name": "k",
-                            "label": "K",
-                            "field": "k",
-                            "align": "center",
-                            "style": "width:48px",
-                        },
-                        {
-                            "name": "augmented_prompt",
-                            "label": "Augmented prompt",
-                            "field": "augmented_prompt",
-                            "align": "left",
-                        },
-                        {
-                            "name": "response_length",
-                            "label": "Response length",
-                            "field": "response_length",
-                            "align": "center",
-                            "style": "width:140px",
-                        },
-                        {
-                            "name": "result",
-                            "label": "Result",
-                            "field": "result",
-                            "align": "center",
-                            "style": "width:100px",
-                        },
-                    ]
-
-                    rows_data = []
-                    for c in candidates:
-                        if c.get("_guardrail_side"):
-                            result_label = "Mitigated"
-                        elif c["error"]:
-                            result_label = "Error"
-                        elif c["is_best"] and is_jailbreak_step:
-                            result_label = "Jailbreak"
-                        elif c["is_best"] and not is_jailbreak_step:
-                            result_label = "Mitigated"
-                        else:
-                            result_label = "—"
-                        aug = c.get("augmented_prompt") or ""
-                        rows_data.append(
-                            {
-                                "k": c["k"],
-                                "augmented_prompt": (aug[:80] + "…")
-                                if len(aug) > 80
-                                else aug or "—",
-                                "response_length": c["response_length"],
-                                "result": result_label,
-                                "_is_best": c["is_best"],
-                                "_full_prompt": aug,
-                                "_response": c.get("response") or "",
-                                "_guardrail_side": c.get("_guardrail_side") or "",
-                                "_guardrail_explanation": c.get(
-                                    "_guardrail_explanation"
-                                )
-                                or "",
-                            }
-                        )
-
-                    tbl = (
-                        ui.table(columns=columns, rows=rows_data, row_key="k")
-                        .classes("w-full text-xs")
-                        .props("dense flat")
-                    )
-                    if detail_mode:
-                        tbl.props(
-                            f":expanded-rows='{json.dumps([r['k'] for r in rows_data])}'"
-                        )
-
-                    tbl.add_slot(
-                        "body",
-                        r"""
-<q-tr :props="props" @click="props.expand = !props.expand"
-      :class="props.row._is_best ? 'bg-grey-2' : ''"
-      style="cursor:pointer">
-  <q-td key="k" :props="props">{{ props.row.k }}</q-td>
-  <q-td key="augmented_prompt" :props="props"
-        style="white-space:pre-wrap;word-break:break-word;max-width:320px">
-    {{ props.row.augmented_prompt }}
-  </q-td>
-  <q-td key="response_length" :props="props">{{ props.row.response_length }}</q-td>
-  <q-td key="result" :props="props">
-    <q-badge v-if="props.row.result === 'Jailbreak'" color="negative" class="text-xs">Jailbreak</q-badge>
-    <q-badge v-else-if="props.row.result === 'Mitigated'" color="positive" class="text-xs">Mitigated</q-badge>
-    <q-badge v-else-if="props.row.result === 'Error'" color="warning" class="text-xs">Error</q-badge>
-    <span v-else class="text-grey-5">&#8212;</span>
-  </q-td>
-</q-tr>
-<q-tr v-show="props.expand" :props="props" @click="props.expand = false" style="cursor:pointer">
-  <q-td colspan="100%" class="bg-grey-1">
-    <div class="q-pa-sm">
-      <div class="row items-center justify-between q-mb-xs">
-        <span class="text-caption text-weight-bold text-uppercase text-grey-6">PROMPT SENT TO TARGET</span>
-        <span :data-cliptext="props.row._full_prompt||''" onclick="event.stopPropagation();var s=this,t=s.dataset.cliptext||'',ic=s.querySelector('.q-icon');if(window.navigator&&window.navigator.clipboard)window.navigator.clipboard.writeText(t);if(ic)ic.textContent='check';s.setAttribute('title','Copied to clipboard');setTimeout(function(){if(ic)ic.textContent='content_copy';s.setAttribute('title','Copy to clipboard');},2000);" style="display:inline-flex;cursor:pointer" title="Copy to clipboard"><q-btn flat dense size="xs" icon="content_copy" color="grey-6" /></span>
-      </div>
-      <pre style="font-size:11px;padding:8px;background:white;border:1px solid #e0e0e0;
-                  border-radius:4px;margin-bottom:8px;white-space:pre-wrap;word-break:break-word">{{ props.row._full_prompt || '\u2014' }}</pre>
-      <template v-if="props.row._guardrail_side !== 'before'">
-        <div class="row items-center justify-between q-mb-xs">
-          <span class="text-caption text-weight-bold text-uppercase text-grey-6">TARGET RESPONSE</span>
-          <span :data-cliptext="props.row._response||''" onclick="event.stopPropagation();var s=this,t=s.dataset.cliptext||'',ic=s.querySelector('.q-icon');if(window.navigator&&window.navigator.clipboard)window.navigator.clipboard.writeText(t);if(ic)ic.textContent='check';s.setAttribute('title','Copied to clipboard');setTimeout(function(){if(ic)ic.textContent='content_copy';s.setAttribute('title','Copy to clipboard');},2000);" style="display:inline-flex;cursor:pointer" title="Copy to clipboard"><q-btn flat dense size="xs" icon="content_copy" color="grey-6" /></span>
-        </div>
-        <pre style="font-size:11px;padding:8px;background:white;border:1px solid #e0e0e0;
-                    border-radius:4px;margin-bottom:8px;white-space:pre-wrap;word-break:break-word">{{ props.row._response || 'No response recorded.' }}</pre>
-      </template>
-      <div v-if="props.row._guardrail_side === 'before'" style="margin-bottom:8px">
-        <div class="text-caption text-weight-bold text-uppercase q-mb-xs" style="color:#c2410c">&#x26a0; BEFORE GUARDRAIL &#x2014; BLOCKED</div>
-        <pre style="font-size:11px;padding:10px;background:#fff7ed;border:2px solid #f97316;border-radius:4px;white-space:pre-wrap;word-break:break-word;margin:0"><span v-if="props.row._guardrail_categories && props.row._guardrail_categories.length" style="font-weight:700;color:#c2410c">Categories: </span><span v-if="props.row._guardrail_categories && props.row._guardrail_categories.length" style="color:#9a3412">{{ props.row._guardrail_categories.join(', ') }}</span><span v-if="props.row._guardrail_categories && props.row._guardrail_categories.length">&#10;&#10;</span><span style="font-weight:700;color:#c2410c">Explanation: </span><span style="color:#6b7280">{{ props.row._guardrail_explanation }}</span></pre>
-      </div>
-      <div v-else-if="props.row._guardrail_side === 'after'" style="margin-bottom:8px">
-        <div class="text-caption text-weight-bold text-uppercase q-mb-xs" style="color:#dc2626">&#x1f6ab; AFTER GUARDRAIL &#x2014; CENSORED</div>
-        <pre style="font-size:11px;padding:10px;background:#fef2f2;border:2px solid #ef4444;border-radius:4px;white-space:pre-wrap;word-break:break-word;margin:0"><span v-if="props.row._guardrail_categories && props.row._guardrail_categories.length" style="font-weight:700;color:#dc2626">Categories: </span><span v-if="props.row._guardrail_categories && props.row._guardrail_categories.length" style="color:#991b1b">{{ props.row._guardrail_categories.join(', ') }}</span><span v-if="props.row._guardrail_categories && props.row._guardrail_categories.length">&#10;&#10;</span><span style="font-weight:700;color:#dc2626">Explanation: </span><span style="color:#6b7280">{{ props.row._guardrail_explanation }}</span></pre>
-      </div>
-      <div v-else-if="props.row._guardrail_side" style="margin-bottom:8px">
-        <div class="text-caption text-weight-bold text-uppercase q-mb-xs" style="color:#616161">&#x1f6e1; GUARDRAIL &#x2014; BLOCKED</div>
-        <pre style="font-size:11px;padding:10px;background:#f5f5f5;border:2px solid #9e9e9e;border-radius:4px;white-space:pre-wrap;word-break:break-word;margin:0"><span v-if="props.row._guardrail_categories && props.row._guardrail_categories.length" style="font-weight:700;color:#616161">Categories: </span><span v-if="props.row._guardrail_categories && props.row._guardrail_categories.length" style="color:#374151">{{ props.row._guardrail_categories.join(', ') }}</span><span v-if="props.row._guardrail_categories && props.row._guardrail_categories.length">&#10;&#10;</span><span style="font-weight:700;color:#616161">Explanation: </span><span style="color:#6b7280">{{ props.row._guardrail_explanation }}</span></pre>
-      </div>
-    </div>
-  </q-td>
-</q-tr>
-""",
-                    )
-
-            if not detail_mode:
-                self._wire_expand_toggle(body_col)
-
-    # ── H4rm3l formatting ─────────────────────────────────────────────────────
-
-    @staticmethod
-    def _format_h4rm3l_program(program: str) -> str:
-        """Convert an h4rm3l program string to a human-readable arrow chain."""
-        if not program or not isinstance(program, str):
-            return program or ""
-        p = program.strip()
-        import re as _re  # noqa: PLC0415
-
-        # Match function/class calls in the chain
-        names = _re.findall(r"\b([A-Za-z][A-Za-z0-9_]*)(?:\s*\()", p)
-        if names:
-            # Skip generic wrappers and chain methods
-            skip = {"Apply", "Compose", "Pipeline", "then"}
-            filtered = [n for n in names if n not in skip]
-            if filtered:
-
-                def _camel_to_words(n: str) -> str:
-                    # Handle acronyms followed by words: DANDecorator → DAN Decorator
-                    s = _re.sub(r"([A-Z]+)([A-Z][a-z])", r"\1 \2", n)
-                    # Handle lowercase/digit followed by uppercase: refusalSuppression → refusal Suppression
-                    s = _re.sub(r"([a-z\d])([A-Z])", r"\1 \2", s)
-                    return s
-
-                return " → ".join(_camel_to_words(n) for n in filtered)
-        # Fallback: title-case snake_case names
-        if "_" in p and " " not in p:
-            return p.replace("_", " ").title()
-        return p
-
-    @staticmethod
-    def _extract_guardrail_from_response(response_value) -> tuple:
-        """Detect a guardrail event embedded in a response value.
-
-        Returns (actual_response, guardrail_side, guardrail_explanation, guardrail_categories) where:
-        - actual_response: real response string/value, or None if before-guardrail
-        - guardrail_side: "before" | "after" | "unknown" | ""
-        - guardrail_explanation: human-readable reason string
-        - guardrail_categories: list of harm category strings
-        """
-        # Legacy string-encoded format produced by old versions of
-        # tap/generation.py: "[GUARDRAIL:<side>] <reasoning>".
-        # New runs store a proper guardrail dict — this branch handles
-        # traces already persisted in the database with the old format.
-        if isinstance(response_value, str):
-            import re as _re
-
-            _m = _re.match(r"^\[GUARDRAIL:(\w+)\]\s*(.*)", response_value, _re.DOTALL)
-            if _m:
-                side = _m.group(1)
-                explanation = _m.group(2).strip() or "Blocked by guardrail"
-                return None, side, explanation, []
-            return response_value, "", "", []
-
-        if not isinstance(response_value, dict):
-            return response_value, "", "", []
-
-        # New format: adapter_type == "guardrail" with agent_specific_data
-        if response_value.get("adapter_type") == "guardrail":
-            info = response_value.get("agent_specific_data") or {}
-            side = info.get("side", "unknown")
-            explanation = str(
-                info.get("reasoning")
-                or info.get("message")
-                or info.get("explanation")
-                or "Blocked by guardrail"
-            )
-            categories = info.get("categories") or []
-            if side == "after":
-                actual = info.get("target_response") or None
-            else:
-                actual = None
-            return actual, side, explanation, categories
-
-        # Legacy format: dict with side key directly (from tracker extraction)
-        if response_value.get("side") in ("before", "after", "unknown"):
-            side = response_value.get("side", "unknown")
-            explanation = str(
-                response_value.get("reasoning")
-                or response_value.get("message")
-                or response_value.get("explanation")
-                or "Blocked by guardrail"
-            )
-            categories = response_value.get("categories") or []
-            if side == "after":
-                actual = response_value.get("target_response") or None
-            else:
-                actual = None
-            return actual, side, explanation, categories
-
-        return response_value, "", "", []
-
-    # ── PAIR ──────────────────────────────────────────────────────────────────
-
-    @staticmethod
-    def _parse_pair_traces(traces: list[dict]) -> list[dict]:
-        """Parse PAIR traces into per-iteration rows.
-
-        Each row:
-          iteration  – 1-based iteration number
-          prompt     – prompt sent to target
-          response   – target response
-          score      – judge score (int or None)
-          is_best    – True if this had the highest score
-        """
-        sorted_traces = sorted(traces, key=lambda x: x.get("sequence", 0))
-        rows: list[dict] = []
-
-        for td in sorted_traces:
-            content = td.get("content")
-            if not isinstance(content, dict):
-                continue
-            step_name = str(content.get("step_name") or "")
-            if "Iteration" not in step_name and "iteration" not in step_name:
-                continue
-            metadata = content.get("metadata") or {}
-            iteration = int(metadata.get("iteration") or len(rows) + 1)
-            req = content.get("request") or {}
-            prompt = req.get("prompt") or "" if isinstance(req, dict) else str(req)
-            if isinstance(prompt, list):
-                user_msgs = [
-                    m.get("content", "") for m in prompt if m.get("role") == "user"
-                ]
-                prompt = user_msgs[-1] if user_msgs else ""
-            resp = content.get("response")
-            resp, _pair_g_side, _pair_g_expl, _pair_g_cats = (
-                DashboardPage._extract_guardrail_from_response(resp)
-            )
-            # Fallback: guardrail info in metadata (older traces)
-            if not _pair_g_side:
-                _gi = metadata.get("guardrail_info") or {}
-                if not _gi:
-                    _tc = metadata.get("target_call") or {}
-                    _gi = _tc.get("guardrail_info") or {}
-                if _gi.get("side"):
-                    _pair_g_side = _gi["side"]
-                    _pair_g_expl = str(
-                        _gi.get("reasoning")
-                        or _gi.get("message")
-                        or _gi.get("explanation")
-                        or "Blocked by guardrail"
-                    )
-                    _pair_g_cats = _gi.get("categories") or []
-            if isinstance(resp, dict):
-                response = (
-                    resp.get("generated_text") or resp.get("completion") or str(resp)
-                )
-            elif resp is not None:
-                response = str(resp)
-            else:
-                response = ""
-            score_raw = (
-                metadata.get("score")
-                or metadata.get("judge_score")
-                or content.get("score")
-            )
-            try:
-                score = int(float(score_raw)) if score_raw is not None else None
-            except (TypeError, ValueError):
-                score = None
-            rows.append(
-                {
-                    "iteration": iteration,
-                    "prompt": str(prompt),
-                    "response": response,
-                    "score": score,
-                    "is_best": False,
-                    "_guardrail_side": _pair_g_side,
-                    "_guardrail_explanation": _pair_g_expl,
-                    "_guardrail_categories": _pair_g_cats,
-                }
-            )
-
-        # Mark best score
-        if rows:
-            scored = [r for r in rows if r["score"] is not None]
-            if scored:
-                best = max(scored, key=lambda r: r["score"])  # type: ignore[arg-type]
-                best["is_best"] = True
-
-        return rows
-
-    def _render_pair_goal_card(
-        self, row: dict, steps: list[dict], detail_mode: bool = False
-    ) -> None:
-        """Render a PAIR goal card as a conversation with per-iteration steps."""
-        with self._goal_card_shell(row, detail_mode):
-            if not steps:
-                ui.label("No PAIR iteration data recorded.").classes(
-                    "text-sm text-grey-6"
-                )
-            else:
-                with ui.column().classes("w-full gap-0 mt-1") as body_col:
-                    if not detail_mode:
-                        body_col.set_visibility(False)
-
-                    for step in steps:
-                        iteration = step["iteration"]
-                        score = step["score"]
-                        is_best = step["is_best"]
-                        prompt = step["prompt"]
-                        response = step["response"]
-                        _guardrail_side = step.get("_guardrail_side") or ""
-                        _guardrail_explanation = (
-                            step.get("_guardrail_explanation") or ""
-                        )
-                        _guardrail_categories = step.get("_guardrail_categories") or []
-
-                        with ui.row().classes("items-center gap-2 mt-3 mb-1 px-1"):
-                            _iter_label = f"Iteration {iteration}"
-                            if score is not None:
-                                _iter_label += f" — Score {score}/10"
-                            if is_best:
-                                _iter_label += " — Best"
-                            ui.label(_iter_label).classes(
-                                "text-xs font-semibold text-grey-6 uppercase tracking-wide"
-                            )
-
-                        with ui.row().classes("w-full items-center justify-between"):
-                            ui.label("PROMPT SENT TO TARGET").classes(
-                                "text-[10px] text-grey-6 font-semibold uppercase tracking-wide px-1"
-                            )
-                            ui.button(icon="content_copy").props(
-                                "flat dense size=xs color=grey-6"
-                            ).tooltip("Copy to clipboard").on(
-                                "click",
-                                js_handler=f"(event) => {{var b=event.currentTarget,ic=b.querySelector('.q-icon');if(navigator.clipboard)navigator.clipboard.writeText({json.dumps(prompt or '')});if(ic){{ic.textContent='check';setTimeout(function(){{ic.textContent='content_copy';}},2000);}}}}",
-                            )
-                        ui.html(
-                            '<pre style="font-size:11px;padding:8px;background:white;border:1px solid #e0e0e0;'
-                            'border-radius:4px;margin-bottom:6px;white-space:pre-wrap;word-break:break-word">'
-                            + html.escape(prompt or "—")
-                            + "</pre>"
-                        )
-
-                        if _guardrail_side == "before":
-                            self._render_guardrail_event_block(
-                                {
-                                    "side": "before",
-                                    "explanation": _guardrail_explanation,
-                                    "categories": _guardrail_categories,
-                                }
-                            )
-                        else:
-                            with ui.row().classes(
-                                "w-full items-center justify-between"
-                            ):
-                                ui.label("TARGET RESPONSE").classes(
-                                    "text-[10px] text-grey-6 font-semibold uppercase tracking-wide px-1"
-                                )
-                                ui.button(icon="content_copy").props(
-                                    "flat dense size=xs color=grey-6"
-                                ).tooltip("Copy to clipboard").on(
-                                    "click",
-                                    js_handler=f"(event) => {{var b=event.currentTarget,ic=b.querySelector('.q-icon');if(navigator.clipboard)navigator.clipboard.writeText({json.dumps(response or '')});if(ic){{ic.textContent='check';setTimeout(function(){{ic.textContent='content_copy';}},2000);}}}}",
-                                )
-                            ui.html(
-                                '<pre style="font-size:11px;padding:8px;background:white;border:1px solid #e0e0e0;'
-                                'border-radius:4px;white-space:pre-wrap;word-break:break-word">'
-                                + html.escape(response or "No response recorded.")
-                                + "</pre>"
-                            )
-                            if _guardrail_side:
-                                self._render_guardrail_event_block(
-                                    {
-                                        "side": _guardrail_side,
-                                        "explanation": _guardrail_explanation,
-                                        "categories": _guardrail_categories,
-                                    }
-                                )
-
-                        if iteration < steps[-1]["iteration"]:
-                            ui.separator().classes("mt-2 mb-0")
-
-                if not detail_mode:
-                    self._wire_expand_toggle(body_col)
-
-    # ── AutoDAN-Turbo ─────────────────────────────────────────────────────────
-
-    @staticmethod
-    def _parse_autodan_traces(traces: list[dict]) -> list[dict]:
-        """Parse AutoDAN-Turbo traces into per-epoch step rows."""
-        sorted_traces = sorted(traces, key=lambda x: x.get("sequence", 0))
-
-        steps: dict[tuple, dict] = {}
-        warmup_summary: dict | None = None
-
-        for td in sorted_traces:
-            content = td.get("content")
-            if not isinstance(content, dict):
-                continue
-            step_name = str(content.get("step_name") or "")
-            # AutoDAN-Turbo stores all payload fields directly in content
-            # (no nested "metadata" dict), so read epoch/iteration/stream
-            # from content directly.
-            metadata = content.get("metadata") or {}
-
-            # phase / subphase are stored directly in content by emit_phase_trace.
-            # step_name is NOT persisted to the backend, so we cannot use it for
-            # phase detection when reading DB traces.
-            phase_raw = str(content.get("phase") or "").upper()
-            subphase_raw = str(content.get("subphase") or "").upper()
-            if not phase_raw:
-                # Legacy fallback: derive from step_name for locally-stored traces
-                phase_raw = "WARMUP" if "warmup" in step_name.lower() else "LIFELONG"
-
-            if phase_raw == "WARMUP" and subphase_raw == "SUMMARIZATION":
-                warmup_summary = {
-                    "phase": "WARMUP_SUMMARY",
-                    "iteration": int(
-                        content.get("iteration") or metadata.get("iteration") or 0
-                    ),
-                    "epoch": -1,
-                    "stream": -1,
-                    "strategy": (content.get("strategy") or metadata.get("strategy")),
-                    "score": None,
-                    "is_best": False,
-                    "generated_prompt": None,
-                    "target_response": None,
-                    "assessment": None,
-                    "score_delta": None,
-                }
-                continue
-
-            # Skip bookend traces — no display content
-            if subphase_raw in ("PHASE_START", "PHASE_END", "SKIP_FINALIZED"):
-                continue
-
-            phase = phase_raw if phase_raw in ("WARMUP", "LIFELONG") else "LIFELONG"
-            # Payload fields are top-level in content; prefer them over the
-            # (usually empty) nested metadata dict.
-            iteration = int(content.get("iteration") or metadata.get("iteration") or 0)
-            epoch = int(content.get("epoch") or metadata.get("epoch") or 0)
-            stream = int(content.get("stream") or metadata.get("stream") or 0)
-            # Use (phase, iteration, epoch, stream) so each inner-loop step
-            # gets its own row while still merging the 3 sub-traces
-            # (Generation / Target Query / Scoring) that share the same key.
-            key = (phase, iteration, epoch, stream)
-
-            if key not in steps:
-                steps[key] = {
-                    "phase": phase,
-                    "iteration": iteration,
-                    "epoch": epoch,
-                    "stream": stream,
-                    "score": None,
-                    "is_best": False,
-                    "generated_prompt": None,
-                    "target_response": None,
-                    "assessment": None,
-                    "strategy": None,
-                    "score_delta": None,
-                    "_guardrail_side": "",
-                    "_guardrail_explanation": "",
-                }
-
-            step = steps[key]
-
-            # ── Attacker / jailbreak prompt ───────────────────────────────
-            # Traces store the final prompt directly as "generated_prompt".
-            # Older/router-style traces may nest it under request["prompt"].
-            if content.get("generated_prompt"):
-                step["generated_prompt"] = str(content["generated_prompt"])
-            else:
-                req = content.get("request") or {}
-                if isinstance(req, dict) and req.get("prompt"):
-                    step["generated_prompt"] = req["prompt"]
-
-            # ── Target response ───────────────────────────────────────────
-            # Traces store it as "target_response"; legacy path uses "response".
-            raw_resp = content.get("target_response") or content.get("response")
-            if raw_resp:
-                raw_resp, _adan_g_side, _adan_g_expl, _adan_g_cats = (
-                    DashboardPage._extract_guardrail_from_response(raw_resp)
-                )
-                if _adan_g_side:
-                    step["_guardrail_side"] = _adan_g_side
-                    step["_guardrail_explanation"] = _adan_g_expl
-                    step["_guardrail_categories"] = _adan_g_cats
-                if raw_resp is not None:
-                    if isinstance(raw_resp, dict):
-                        step["target_response"] = (
-                            raw_resp.get("generated_text")
-                            or raw_resp.get("completion")
-                            or str(raw_resp)
-                        )
-                    else:
-                        step["target_response"] = str(raw_resp)
-
-            # ── Score / assessment / strategy ─────────────────────────────
-            score_raw = content.get("score") or metadata.get("judge_score")
-            if score_raw is not None:
-                try:
-                    step["score"] = float(score_raw)
-                except (TypeError, ValueError):
-                    pass
-
-            if content.get("assessment"):
-                step["assessment"] = str(content["assessment"])
-            if content.get("strategy"):
-                step["strategy"] = content["strategy"]
-
-            score_delta_raw = content.get("score_delta") or metadata.get("score_delta")
-            if score_delta_raw is not None:
-                try:
-                    step["score_delta"] = float(score_delta_raw)
-                except (TypeError, ValueError):
-                    pass
-
-        _phase_order = {"WARMUP": 0, "LIFELONG": 1}
-        result = [
-            steps[k]
-            for k in sorted(
-                steps,
-                key=lambda t: (_phase_order.get(t[0], 9), t[1], t[2], t[3]),
-            )
-        ]
-        for phase_label in ("WARMUP", "LIFELONG"):
-            phase_steps = [s for s in result if s["phase"] == phase_label]
-            scored = [s for s in phase_steps if s["score"] is not None]
-            if scored:
-                best = max(scored, key=lambda s: s["score"])  # type: ignore[arg-type]
-                best["is_best"] = True
-
-        if warmup_summary:
-            last_warmup_idx = -1
-            for i, s in enumerate(result):
-                if s["phase"] == "WARMUP":
-                    last_warmup_idx = i
-            result.insert(last_warmup_idx + 1, warmup_summary)
-
-        return result
-
-    def _render_autodan_goal_card(
-        self, row: dict, steps: list[dict], detail_mode: bool = False
-    ) -> None:
-        """Render AutoDAN-Turbo goal card as a phase-divided conversation."""
-        with self._goal_card_shell(row, detail_mode):
-            if not steps:
-                ui.label("No AutoDAN-Turbo trace data recorded.").classes(
-                    "text-sm text-grey-6"
-                )
-                return
-
-            with ui.column().classes("w-full gap-3 mt-2") as body_col:
-                if not detail_mode:
-                    body_col.set_visibility(False)
-
-                phase_groups: list[tuple[str, list[dict]]] = []
-                for step in steps:
-                    phase = step["phase"]
-                    display_phase = (
-                        "WARMUP"
-                        if phase in ("WARMUP", "WARMUP_SUMMARY")
-                        else "LIFELONG"
-                    )
-                    if not phase_groups or phase_groups[-1][0] != display_phase:
-                        phase_groups.append((display_phase, []))
-                    phase_groups[-1][1].append(step)
-
-                for display_phase, phase_steps in phase_groups:
-                    _is_warmup = display_phase == "WARMUP"
-                    _phase_border = (
-                        "border-blue-grey-3" if _is_warmup else "border-teal-3"
-                    )
-                    _phase_header_bg = (
-                        "background:#eceff1" if _is_warmup else "background:#e0f2f1"
-                    )
-                    _phase_label_text = "Warm-Up" if _is_warmup else "Lifelong"
-                    _phase_icon = "explore" if _is_warmup else "loop"
-
-                    with ui.card().tight().classes(f"w-full border {_phase_border}"):
-                        with (
-                            ui.row()
-                            .classes("items-center gap-2 px-3 py-2 w-full")
-                            .style(_phase_header_bg)
-                        ):
-                            ui.icon(_phase_icon, size="xs").classes("text-grey-7")
-                            ui.label(_phase_label_text).classes(
-                                "text-xs font-bold text-grey-8 uppercase tracking-widest"
-                            )
-
-                        with ui.column().classes("w-full gap-2 p-2"):
-                            # Split WARMUP_SUMMARY out first, then group the
-                            # remaining steps by iteration so each outer-loop
-                            # iteration gets a single collapsible header with
-                            # its epoch cards nested inside.
-                            _summary_steps = [
-                                s for s in phase_steps if s["phase"] == "WARMUP_SUMMARY"
-                            ]
-                            _iter_steps = [
-                                s for s in phase_steps if s["phase"] != "WARMUP_SUMMARY"
-                            ]
-
-                            # Build ordered iteration groups preserving sort order
-                            _iter_groups: list[tuple[int, list[dict]]] = []
-                            for _step in _iter_steps:
-                                _it = _step.get("iteration", 0)
-                                if not _iter_groups or _iter_groups[-1][0] != _it:
-                                    _iter_groups.append((_it, []))
-                                _iter_groups[-1][1].append(_step)
-
-                            for _iter_num, _epoch_steps in _iter_groups:
-                                # Iteration sub-header (only shown when there
-                                # are multiple iterations in this phase)
-                                _iter_has_best = any(
-                                    s.get("is_best") for s in _epoch_steps
-                                )
-                                _iter_best_score = max(
-                                    (
-                                        s["score"]
-                                        for s in _epoch_steps
-                                        if s.get("score") is not None
-                                    ),
-                                    default=None,
-                                )
-                                if len(_iter_groups) > 1:
-                                    with ui.row().classes(
-                                        "items-center gap-2 px-1 py-0.5"
-                                    ):
-                                        ui.label(f"Iteration {_iter_num + 1}").classes(
-                                            "text-[11px] font-bold text-grey-6 uppercase tracking-widest"
-                                        )
-                                        if _iter_has_best:
-                                            _ib_str = (
-                                                f"  best {_iter_best_score:.1f} / 10"
-                                                if _iter_best_score is not None
-                                                else ""
-                                            )
-                                            ui.badge(
-                                                f"Best{_ib_str}", color="positive"
-                                            ).classes("text-xs")
-
-                                for step in _epoch_steps:
-                                    score = step["score"]
-                                    is_best = step["is_best"]
-                                    generated_prompt = step["generated_prompt"]
-                                    target_response = step["target_response"]
-                                    assessment = step.get("assessment") or ""
-                                    strategy = step.get("strategy")
-                                    score_delta = step.get("score_delta")
-                                    _adan_g_side = step.get("_guardrail_side") or ""
-                                    _adan_g_expl = (
-                                        step.get("_guardrail_explanation") or ""
-                                    )
-                                    _epoch_border = (
-                                        "border-positive"
-                                        if is_best
-                                        else "border-grey-3"
-                                    )
-
-                                    with (
-                                        ui.card()
-                                        .tight()
-                                        .classes(f"w-full border {_epoch_border}")
-                                    ):
-                                        with (
-                                            ui.row()
-                                            .classes(
-                                                "items-center gap-2 px-3 py-1.5 w-full"
-                                            )
-                                            .style("background:#f5f5f5")
-                                        ):
-                                            _score_str = (
-                                                f" — score {score:.1f} / 10"
-                                                if score is not None
-                                                else ""
-                                            )
-                                            _ep = step.get("epoch", 0)
-                                            # Only show "Epoch N" when there
-                                            # are multiple epochs per iteration
-                                            _ep_count = max(
-                                                (
-                                                    s.get("epoch", 0)
-                                                    for s in _epoch_steps
-                                                ),
-                                                default=0,
-                                            )
-                                            if _ep_count > 0:
-                                                _step_label = (
-                                                    f"Epoch {_ep + 1}{_score_str}"
-                                                )
-                                            else:
-                                                _step_label = f"Iteration {_iter_num + 1}{_score_str}"
-                                            ui.label(_step_label).classes(
-                                                "text-xs font-semibold text-grey-7 uppercase tracking-wide"
-                                            )
-                                            if is_best:
-                                                ui.badge(
-                                                    "Best", color="positive"
-                                                ).classes("text-xs")
-                                            ui.space()
-
-                                        with ui.column().classes("p-3 gap-2"):
-                                            with ui.row().classes(
-                                                "w-full items-center justify-between"
-                                            ):
-                                                ui.label("Attacker").classes(
-                                                    "text-[10px] font-semibold text-grey-5 uppercase tracking-wide"
-                                                )
-                                                ui.button(icon="content_copy").props(
-                                                    "flat dense size=xs color=grey-6"
-                                                ).tooltip("Copy to clipboard").on(
-                                                    "click",
-                                                    js_handler=f"(event) => {{var b=event.currentTarget,ic=b.querySelector('.q-icon');if(navigator.clipboard)navigator.clipboard.writeText({json.dumps(generated_prompt or '')});if(ic){{ic.textContent='check';setTimeout(function(){{ic.textContent='content_copy';}},2000);}}}}",
-                                                )
-                                            ui.html(
-                                                '<pre style="font-size:11px;padding:8px;background:white;'
-                                                "border:1px solid #e0e0e0;border-radius:4px;margin:0;"
-                                                'white-space:pre-wrap;word-break:break-word">'
-                                                + html.escape(generated_prompt or "—")
-                                                + "</pre>"
-                                            )
-
-                                            if _adan_g_side == "before":
-                                                self._render_guardrail_event_block(
-                                                    {
-                                                        "side": "before",
-                                                        "explanation": _adan_g_expl,
-                                                    }
-                                                )
-                                            else:
-                                                with ui.row().classes(
-                                                    "w-full items-center justify-between"
-                                                ):
-                                                    ui.label("Target response").classes(
-                                                        "text-[10px] font-semibold text-grey-5 uppercase tracking-wide"
-                                                    )
-                                                    ui.button(
-                                                        icon="content_copy"
-                                                    ).props(
-                                                        "flat dense size=xs color=grey-6"
-                                                    ).tooltip("Copy to clipboard").on(
-                                                        "click",
-                                                        js_handler=f"(event) => {{var b=event.currentTarget,ic=b.querySelector('.q-icon');if(navigator.clipboard)navigator.clipboard.writeText({json.dumps(target_response or '')});if(ic){{ic.textContent='check';setTimeout(function(){{ic.textContent='content_copy';}},2000);}}}}",
-                                                    )
-                                                ui.html(
-                                                    '<pre style="font-size:11px;padding:8px;background:white;'
-                                                    "border:1px solid #e0e0e0;border-radius:4px;margin:0;"
-                                                    'white-space:pre-wrap;word-break:break-word">'
-                                                    + html.escape(
-                                                        target_response
-                                                        or "No response recorded."
-                                                    )
-                                                    + "</pre>"
-                                                )
-                                                if _adan_g_side:
-                                                    self._render_guardrail_event_block(
-                                                        {
-                                                            "side": _adan_g_side,
-                                                            "explanation": _adan_g_expl,
-                                                        }
-                                                    )
-
-                                            if assessment:
-                                                with ui.row().classes(
-                                                    "w-full items-center justify-between"
-                                                ):
-                                                    ui.label(
-                                                        "Scorer assessment"
-                                                    ).classes(
-                                                        "text-[10px] font-semibold text-grey-5 uppercase tracking-wide"
-                                                    )
-                                                    ui.button(
-                                                        icon="content_copy"
-                                                    ).props(
-                                                        "flat dense size=xs color=grey-6"
-                                                    ).tooltip("Copy to clipboard").on(
-                                                        "click",
-                                                        js_handler=f"(event) => {{var b=event.currentTarget,ic=b.querySelector('.q-icon');if(navigator.clipboard)navigator.clipboard.writeText({json.dumps(assessment or '')});if(ic){{ic.textContent='check';setTimeout(function(){{ic.textContent='content_copy';}},2000);}}}}",
-                                                    )
-                                                ui.html(
-                                                    '<pre style="font-size:11px;padding:8px;background:#fff8e1;'
-                                                    "border:1px solid #ffe082;border-radius:4px;margin:0;"
-                                                    'white-space:pre-wrap;word-break:break-word">'
-                                                    + html.escape(assessment)
-                                                    + "</pre>"
-                                                )
-
-                                            if strategy is not None and isinstance(
-                                                strategy, dict
-                                            ):
-                                                s_name = strategy.get("Strategy")
-                                                s_defn = strategy.get("Definition")
-                                                if s_name or s_defn:
-                                                    _delta_str = (
-                                                        f" (+{score_delta:.1f})"
-                                                        if score_delta
-                                                        else ""
-                                                    )
-                                                    _strat_text = ""
-                                                    if s_name:
-                                                        _strat_text += (
-                                                            f"Strategy: {s_name}\n"
-                                                        )
-                                                    if s_defn:
-                                                        _strat_text += (
-                                                            f"Definition: {s_defn}"
-                                                        )
-                                                    with ui.row().classes(
-                                                        "w-full items-center justify-between"
-                                                    ):
-                                                        ui.label(
-                                                            f"New strategy{_delta_str}"
-                                                        ).classes(
-                                                            "text-[10px] font-semibold text-indigo-6 uppercase tracking-wide"
-                                                        )
-                                                        ui.button(
-                                                            icon="content_copy"
-                                                        ).props(
-                                                            "flat dense size=xs color=grey-6"
-                                                        ).tooltip(
-                                                            "Copy to clipboard"
-                                                        ).on(
-                                                            "click",
-                                                            js_handler=f"(event) => {{var b=event.currentTarget,ic=b.querySelector('.q-icon');if(navigator.clipboard)navigator.clipboard.writeText({json.dumps(_strat_text.strip())});if(ic){{ic.textContent='check';setTimeout(function(){{ic.textContent='content_copy';}},2000);}}}}",
-                                                        )
-                                                    ui.html(
-                                                        '<pre style="font-size:11px;padding:8px;background:#f3f4fd;'
-                                                        "border:1px solid #c5cae9;border-radius:4px;margin:0;"
-                                                        'white-space:pre-wrap;word-break:break-word">'
-                                                        + html.escape(
-                                                            _strat_text.strip()
-                                                        )
-                                                        + "</pre>"
-                                                    )
-
-                            # Render WARMUP_SUMMARY cards (strategy extracted
-                            # by the summarizer after the warmup loop)
-                            for _ws in _summary_steps:
-                                _ws_strategy = _ws.get("strategy") or {}
-                                _ws_name = (
-                                    _ws_strategy.get("Strategy")
-                                    if isinstance(_ws_strategy, dict)
-                                    else None
-                                )
-                                _ws_defn = (
-                                    _ws_strategy.get("Definition")
-                                    if isinstance(_ws_strategy, dict)
-                                    else None
-                                )
-                                if _ws_name or _ws_defn:
-                                    with (
-                                        ui.card()
-                                        .tight()
-                                        .classes("w-full border border-indigo-2")
-                                    ):
-                                        with (
-                                            ui.row()
-                                            .classes("items-center gap-2 px-3 py-2")
-                                            .style("background:#e8eaf6")
-                                        ):
-                                            ui.icon("summarize", size="xs").classes(
-                                                "text-indigo-6"
-                                            )
-                                            ui.label(
-                                                "Summarizer — Strategy Extracted"
-                                            ).classes(
-                                                "text-xs font-bold text-indigo-8 uppercase tracking-widest"
-                                            )
-                                        with ui.column().classes("p-3 gap-1"):
-                                            _strat_text = ""
-                                            if _ws_name:
-                                                _strat_text += f"Strategy: {_ws_name}\n"
-                                            if _ws_defn:
-                                                _strat_text += f"Definition: {_ws_defn}"
-                                            with ui.row().classes(
-                                                "w-full items-center justify-between"
-                                            ):
-                                                ui.label("Extracted strategy").classes(
-                                                    "text-[10px] font-semibold text-indigo-6 uppercase tracking-wide"
-                                                )
-                                                ui.button(icon="content_copy").props(
-                                                    "flat dense size=xs color=grey-6"
-                                                ).tooltip("Copy to clipboard").on(
-                                                    "click",
-                                                    js_handler=f"(event) => {{var b=event.currentTarget,ic=b.querySelector('.q-icon');if(navigator.clipboard)navigator.clipboard.writeText({json.dumps(_strat_text.strip())});if(ic){{ic.textContent='check';setTimeout(function(){{ic.textContent='content_copy';}},2000);}}}}",
-                                                )
-                                            ui.html(
-                                                '<pre style="font-size:11px;padding:8px;background:white;'
-                                                "border:1px solid #c5cae9;border-radius:4px;margin:0;"
-                                                'white-space:pre-wrap;word-break:break-word">'
-                                                + html.escape(_strat_text.strip())
-                                                + "</pre>"
-                                            )
-
-            if not detail_mode:
-                self._wire_expand_toggle(body_col)
-
-    # ── AdvPrefix ─────────────────────────────────────────────────────────────
-
-    @staticmethod
-    def _parse_advprefix_traces(
-        traces: list[dict],
-    ) -> tuple[list[dict], dict]:
-        """Parse AdvPrefix traces into per-prefix rows grouped by meta_prefix."""
-        sorted_traces = sorted(traces, key=lambda x: x.get("sequence", 0))
-
-        gen_stats: dict = {
-            "raw_generated": 0,
-            "after_phase1": 0,
-            "after_phase2": 0,
-        }
-
-        candidates: dict[str, dict] = {}
-
-        for td in sorted_traces:
-            content = td.get("content")
-            if not isinstance(content, dict):
-                continue
-            if "candidates" not in content and "raw_generated" not in content:
-                continue
-            gen_stats["raw_generated"] += int(content.get("raw_generated") or 0)
-            gen_stats["after_phase1"] += int(content.get("after_phase1_filtering") or 0)
-            gen_stats["after_phase2"] += int(content.get("after_phase2_filtering") or 0)
-            for cand in content.get("candidates") or []:
-                if not isinstance(cand, dict):
-                    continue
-                prefix_text = str(cand.get("prefix") or "")
-                if not prefix_text:
-                    continue
-                if prefix_text not in candidates:
-                    _raw_mp = str(cand.get("meta_prefix") or "")
-                    _mp_parts = [p.strip() for p in _raw_mp.split(",") if p.strip()]
-                    _seen: set[str] = set()
-                    _mp_dedup: list[str] = []
-                    for _p in _mp_parts:
-                        if _p not in _seen:
-                            _seen.add(_p)
-                            _mp_dedup.append(_p)
-                    _meta_prefix_str = ", ".join(_mp_dedup)
-                    candidates[prefix_text] = {
-                        "prefix": prefix_text,
-                        "_meta_prefix": _meta_prefix_str,
-                        "_nll": cand.get("prefix_nll"),
-                        "completion": "",
-                        "_bucket": "pending",
-                        "result": "Pending",
-                        "_filtered": "Pending",
-                        "_error": "",
-                    }
-
-        completion_by_prefix: dict[str, dict] = {}
-
-        for td in sorted_traces:
-            content = td.get("content")
-            if not isinstance(content, dict):
-                continue
-            if str(content.get("step_name") or "") != "Target Completion":
-                continue
-
-            metadata = content.get("metadata") or {}
-            full_prefix = str(metadata.get("prefix") or "")
-            lookup_key = full_prefix[:300]
-            error_msg = metadata.get("error_message")
-            surrogate = str(metadata.get("surrogate_attack_prompt") or "")
-
-            response = content.get("response")
-            response, _adv_g_side, _adv_g_expl, _adv_g_cats = (
-                DashboardPage._extract_guardrail_from_response(response)
-            )
-            if isinstance(response, dict):
-                completion = (
-                    response.get("generated_text") or response.get("completion") or ""
-                )
-            elif isinstance(response, str):
-                completion = response
-            elif response is None:
-                completion = ""
-            else:
-                completion = str(response)
-
-            # error_message may carry the guardrail block description when the
-            # guardrail info was not embedded in the response object itself
-            # (e.g. for runs recorded before guardrail_event propagation was added).
-            if not _adv_g_side and error_msg:
-                _emsg_lower = str(error_msg).lower()
-                if (
-                    "before_guardrail" in _emsg_lower
-                    or "before guardrail" in _emsg_lower
-                ):
-                    _adv_g_side = "before"
-                    _adv_g_expl = str(error_msg)
-                    error_msg = None
-                elif (
-                    "after_guardrail" in _emsg_lower or "after guardrail" in _emsg_lower
-                ):
-                    _adv_g_side = "after"
-                    _adv_g_expl = str(error_msg)
-                    error_msg = None
-
-            if _adv_g_side:
-                bucket = "mitigated"
-                result_label = "Mitigated"
-                error_msg = None  # guardrail takes precedence over any error_msg
-            elif error_msg:
-                bucket = "error"
-                result_label = "Error"
-            elif completion:
-                bucket = "mitigated"
-                result_label = "Mitigated"
-            else:
-                bucket = "error"
-                result_label = "Error"
-
-            comp_data = {
-                "prefix": full_prefix,
-                "completion": completion,
-                "_bucket": bucket,
-                "result": result_label,
-                "_filtered": "No",
-                "_error": str(error_msg) if error_msg else "",
-                "_meta_prefix": "",
-                "_surrogate": surrogate,
-                "_guardrail_side": _adv_g_side,
-                "_guardrail_explanation": _adv_g_expl,
-            }
-            completion_by_prefix[lookup_key] = comp_data
-
-            if lookup_key not in candidates:
-                candidates[lookup_key] = {
-                    "prefix": full_prefix,
-                    "_meta_prefix": "",
-                    "_nll": None,
-                    "completion": "",
-                    "_bucket": "pending",
-                    "result": "Pending",
-                    "_filtered": "Pending",
-                    "_error": "",
-                    "_surrogate": surrogate,
-                }
-
-        rows: list[dict] = []
-        for key, cand in candidates.items():
-            comp = completion_by_prefix.get(key)
-            if comp:
-                cand["prefix"] = comp["prefix"]
-                cand["completion"] = comp["completion"]
-                cand["_bucket"] = comp["_bucket"]
-                cand["result"] = comp["result"]
-                cand["_filtered"] = comp["_filtered"]
-                cand["_error"] = comp["_error"]
-                cand["_surrogate"] = comp.get("_surrogate", "")
-                if not cand.get("_meta_prefix") and comp["_meta_prefix"]:
-                    cand["_meta_prefix"] = comp["_meta_prefix"]
-                cand["_guardrail_side"] = comp.get("_guardrail_side") or ""
-                cand["_guardrail_explanation"] = (
-                    comp.get("_guardrail_explanation") or ""
-                )
-            _surrogate = cand.get("_surrogate") or ""
-            _prefix = cand["prefix"]
-            if _surrogate:
-                if "{prefix}" in _surrogate:
-                    cand["_sent_prompt"] = _surrogate.format(prefix=_prefix)
-                else:
-                    cand["_sent_prompt"] = _prefix + " " + _surrogate
-            else:
-                cand["_sent_prompt"] = _prefix
-            rows.append(cand)
-
-        rows.sort(key=lambda r: (r["_meta_prefix"], r["prefix"][:40]))
-        for i, r in enumerate(rows):
-            r["num"] = i + 1
-
-        unmatched_jailbreaks = 0
-        for td in sorted_traces:
-            content = td.get("content")
-            if not isinstance(content, dict):
-                continue
-            if str(content.get("step_name") or "") != "Evaluation":
-                continue
-            # Skip coordinator summary traces — they are goal-level aggregates,
-            # not per-prefix evaluation signals.
-            if str(content.get("evaluator") or "") == "tracking_coordinator":
-                continue
-            _result_val = content.get("result")
-            is_success = (
-                content.get("success") is True
-                or content.get("is_success") is True
-                or (
-                    isinstance(_result_val, dict) and _result_val.get("success") is True
-                )
-                or (content.get("score") or 0) > 0
-            )
-            if not is_success:
-                continue
-            meta = content.get("metadata") or {}
-            eval_prefix = str(meta.get("prefix") or "")
-            if eval_prefix:
-                eval_key = eval_prefix[:300]
-                matched = False
-                for r in rows:
-                    if r["prefix"][:300] == eval_key:
-                        r["_bucket"] = "jailbreak"
-                        r["result"] = "Jailbreak"
-                        matched = True
-                if not matched:
-                    unmatched_jailbreaks += 1
-            else:
-                unmatched_jailbreaks += 1
-
-        if unmatched_jailbreaks:
-            marked = 0
-            for r in rows:
-                if marked >= unmatched_jailbreaks:
-                    break
-                if r["_bucket"] in ("mitigated", "error") and not r.get(
-                    "_guardrail_side"
-                ):
-                    r["_bucket"] = "jailbreak"
-                    r["result"] = "Jailbreak"
-                    marked += 1
-
-        return rows, gen_stats
-
-    def _render_advprefix_goal_card(
-        self,
-        row: dict,
-        prefix_rows: list[dict],
-        gen_stats: dict,
-        detail_mode: bool = False,
-    ) -> None:
-        """Render an AdvPrefix goal card as a single flat table."""
-        n_jailbreaks = sum(1 for r in prefix_rows if r["_bucket"] == "jailbreak")
-        n_mitigated = sum(1 for r in prefix_rows if r["_bucket"] == "mitigated")
-        n_errors = sum(1 for r in prefix_rows if r["_bucket"] == "error")
-        n_pending = sum(1 for r in prefix_rows if r["_bucket"] == "pending")
-
-        with self._goal_card_shell(row, detail_mode):
-            if not prefix_rows:
-                ui.label("No AdvPrefix completion data recorded.").classes(
-                    "text-sm text-grey-6"
-                )
-            else:
-                with ui.column().classes("w-full gap-2 mt-1") as body_col:
-                    if not detail_mode:
-                        body_col.set_visibility(False)
-
-                    _raw = gen_stats.get("raw_generated", 0)
-                    _p1 = gen_stats.get("after_phase1", 0)
-                    _p2 = gen_stats.get("after_phase2", 0)
-                    if _raw > 0:
-                        with ui.row().classes(
-                            "items-center gap-1 text-[10px] text-grey-5 mb-1"
-                        ):
-                            ui.label(f"{_raw} generated").classes("font-mono")
-                            ui.label("→").classes("text-grey-4")
-                            ui.label(f"{_p1} after pattern filter").classes("font-mono")
-                            ui.label("→").classes("text-grey-4")
-                            ui.label(f"{_p2} after CE + top-k").classes(
-                                "font-mono font-semibold text-grey-7"
-                            )
-
-                    columns = [
-                        {
-                            "name": "num",
-                            "label": "#",
-                            "field": "num",
-                            "align": "center",
-                            "style": "width:36px",
-                        },
-                        {
-                            "name": "meta_prefix",
-                            "label": "Meta Prefix",
-                            "field": "meta_prefix",
-                            "align": "left",
-                        },
-                        {
-                            "name": "result",
-                            "label": "Result",
-                            "field": "result",
-                            "align": "center",
-                            "style": "width:100px",
-                        },
-                    ]
-
-                    tbl_rows = [
-                        {
-                            "num": r["num"],
-                            "meta_prefix": r.get("_meta_prefix") or "—",
-                            "result": r["result"],
-                            "_bucket": r["_bucket"],
-                            "_full_sent_prompt": r.get("_sent_prompt") or r["prefix"],
-                            "_full_prefix": r["prefix"],
-                            "_completion": r.get("completion") or "",
-                            "_error": r.get("_error") or "",
-                            "_guardrail_side": r.get("_guardrail_side") or "",
-                            "_guardrail_explanation": r.get("_guardrail_explanation")
-                            or "",
-                        }
-                        for r in prefix_rows
-                    ]
-
-                    tbl = (
-                        ui.table(columns=columns, rows=tbl_rows, row_key="num")
-                        .classes("w-full text-xs")
-                        .props("dense flat")
-                    )
-                    if detail_mode:
-                        tbl.props(
-                            f":expanded-rows='{json.dumps([r['num'] for r in tbl_rows])}'"
-                        )
-
-                    tbl.add_slot(
-                        "body",
-                        r"""
-<q-tr :props="props" @click="props.expand = !props.expand" style="cursor:pointer">
-  <q-td key="num" :props="props" style="font-size:10px;color:#9e9e9e">{{ props.row.num }}</q-td>
-  <q-td key="meta_prefix" :props="props"
-        style="white-space:nowrap;overflow:hidden;text-overflow:ellipsis;max-width:200px">
-    {{ props.row.meta_prefix }}
-  </q-td>
-  <q-td key="result" :props="props">
-    <q-badge v-if="props.row._bucket === 'jailbreak'" color="negative" class="text-xs">Jailbreak</q-badge>
-    <q-badge v-else-if="props.row._bucket === 'mitigated'" color="positive" class="text-xs">Mitigated</q-badge>
-    <q-badge v-else-if="props.row._bucket === 'pending'" color="grey" class="text-xs">Pending</q-badge>
-    <q-badge v-else color="warning" class="text-xs">Error</q-badge>
-  </q-td>
-</q-tr>
-<q-tr v-show="props.expand" :props="props" @click="props.expand = false" style="cursor:pointer">
-  <q-td colspan="100%" class="bg-grey-1">
-    <div class="q-pa-sm">
-      <template v-if="props.row._full_prefix !== props.row._full_sent_prompt">
-        <div class="row items-center justify-between q-mb-xs">
-          <span class="text-caption text-weight-bold text-uppercase text-grey-6">RAW ADVERSARIAL PREFIX</span>
-          <span :data-cliptext="props.row._full_prefix||''" onclick="event.stopPropagation();var s=this,t=s.dataset.cliptext||'',ic=s.querySelector('.q-icon');if(window.navigator&&window.navigator.clipboard)window.navigator.clipboard.writeText(t);if(ic)ic.textContent='check';s.setAttribute('title','Copied to clipboard');setTimeout(function(){if(ic)ic.textContent='content_copy';s.setAttribute('title','Copy to clipboard');},2000);" style="display:inline-flex;cursor:pointer" title="Copy to clipboard"><q-btn flat dense size="xs" icon="content_copy" color="grey-6" /></span>
-        </div>
-        <pre style="font-size:11px;padding:8px;background:white;border:1px solid #e0e0e0;
-                    border-radius:4px;margin-bottom:8px;white-space:pre-wrap;word-break:break-word">{{ props.row._full_prefix || '—' }}</pre>
-        <div class="row items-center justify-between q-mb-xs">
-          <span class="text-caption text-weight-bold text-uppercase text-grey-6">PROMPT SENT TO TARGET</span>
-          <span :data-cliptext="props.row._full_sent_prompt||''" onclick="event.stopPropagation();var s=this,t=s.dataset.cliptext||'',ic=s.querySelector('.q-icon');if(window.navigator&&window.navigator.clipboard)window.navigator.clipboard.writeText(t);if(ic)ic.textContent='check';s.setAttribute('title','Copied to clipboard');setTimeout(function(){if(ic)ic.textContent='content_copy';s.setAttribute('title','Copy to clipboard');},2000);" style="display:inline-flex;cursor:pointer" title="Copy to clipboard"><q-btn flat dense size="xs" icon="content_copy" color="grey-6" /></span>
-        </div>
-        <pre style="font-size:11px;padding:8px;background:#fff8e1;border:1px solid #ffe082;
-                    border-radius:4px;margin-bottom:8px;white-space:pre-wrap;word-break:break-word">{{ props.row._full_sent_prompt }}</pre>
-      </template>
-      <template v-else>
-        <div class="row items-center justify-between q-mb-xs">
-          <span class="text-caption text-weight-bold text-uppercase text-grey-6">PROMPT SENT TO TARGET</span>
-          <span :data-cliptext="props.row._full_sent_prompt||''" onclick="event.stopPropagation();var s=this,t=s.dataset.cliptext||'',ic=s.querySelector('.q-icon');if(window.navigator&&window.navigator.clipboard)window.navigator.clipboard.writeText(t);if(ic)ic.textContent='check';s.setAttribute('title','Copied to clipboard');setTimeout(function(){if(ic)ic.textContent='content_copy';s.setAttribute('title','Copy to clipboard');},2000);" style="display:inline-flex;cursor:pointer" title="Copy to clipboard"><q-btn flat dense size="xs" icon="content_copy" color="grey-6" /></span>
-        </div>
-        <pre style="font-size:11px;padding:8px;background:white;border:1px solid #e0e0e0;
-                    border-radius:4px;margin-bottom:8px;white-space:pre-wrap;word-break:break-word">{{ props.row._full_sent_prompt || '—' }}</pre>
-      </template>
-      <template v-if="props.row._bucket !== 'pending' && props.row._guardrail_side !== 'before'">
-        <div class="row items-center justify-between q-mb-xs">
-          <span class="text-caption text-weight-bold text-uppercase text-grey-6">TARGET COMPLETION</span>
-          <span :data-cliptext="props.row._completion||''" onclick="event.stopPropagation();var s=this,t=s.dataset.cliptext||'',ic=s.querySelector('.q-icon');if(window.navigator&&window.navigator.clipboard)window.navigator.clipboard.writeText(t);if(ic)ic.textContent='check';s.setAttribute('title','Copied to clipboard');setTimeout(function(){if(ic)ic.textContent='content_copy';s.setAttribute('title','Copy to clipboard');},2000);" style="display:inline-flex;cursor:pointer" title="Copy to clipboard"><q-btn flat dense size="xs" icon="content_copy" color="grey-6" /></span>
-        </div>
-        <pre style="font-size:11px;padding:8px;background:white;border:1px solid #e0e0e0;
-                    border-radius:4px;white-space:pre-wrap;word-break:break-word">{{ props.row._completion || 'No completion recorded.' }}</pre>
-        <div v-if="props.row._error" class="text-caption text-negative text-italic q-mt-xs">
-          Error: {{ props.row._error }}
-        </div>
-      </template>
-      <div v-else-if="props.row._bucket === 'pending'" class="text-caption text-grey-6 text-italic q-mt-xs">
-        This candidate was not executed against the target model.
-      </div>
-      <div v-if="props.row._guardrail_side === 'before'" style="margin-top:4px;margin-bottom:8px">
-        <div class="text-caption text-weight-bold text-uppercase q-mb-xs" style="color:#c2410c">⚠ BEFORE GUARDRAIL — BLOCKED</div>
-        <pre style="font-size:11px;padding:10px;background:#fff7ed;border:2px solid #f97316;border-radius:4px;white-space:pre-wrap;word-break:break-word;margin:0"><span v-if="props.row._guardrail_categories && props.row._guardrail_categories.length" style="font-weight:700;color:#c2410c">Categories: </span><span v-if="props.row._guardrail_categories && props.row._guardrail_categories.length" style="color:#9a3412">{{ props.row._guardrail_categories.join(', ') }}</span><span v-if="props.row._guardrail_categories && props.row._guardrail_categories.length">&#10;&#10;</span><span style="font-weight:700;color:#c2410c">Explanation: </span><span style="color:#6b7280">{{ props.row._guardrail_explanation }}</span></pre>
-      </div>
-      <div v-else-if="props.row._guardrail_side === 'after'" style="margin-top:4px;margin-bottom:8px">
-        <div class="text-caption text-weight-bold text-uppercase q-mb-xs" style="color:#dc2626">🚫 AFTER GUARDRAIL — CENSORED</div>
-        <pre style="font-size:11px;padding:10px;background:#fef2f2;border:2px solid #ef4444;border-radius:4px;white-space:pre-wrap;word-break:break-word;margin:0"><span v-if="props.row._guardrail_categories && props.row._guardrail_categories.length" style="font-weight:700;color:#dc2626">Categories: </span><span v-if="props.row._guardrail_categories && props.row._guardrail_categories.length" style="color:#991b1b">{{ props.row._guardrail_categories.join(', ') }}</span><span v-if="props.row._guardrail_categories && props.row._guardrail_categories.length">&#10;&#10;</span><span style="font-weight:700;color:#dc2626">Explanation: </span><span style="color:#6b7280">{{ props.row._guardrail_explanation }}</span></pre>
-      </div>
-      <div v-else-if="props.row._guardrail_side" style="margin-top:4px;margin-bottom:8px">
-        <div class="text-caption text-weight-bold text-uppercase q-mb-xs" style="color:#616161">🛡 GUARDRAIL — BLOCKED</div>
-        <pre style="font-size:11px;padding:10px;background:#f5f5f5;border:2px solid #9e9e9e;border-radius:4px;white-space:pre-wrap;word-break:break-word;margin:0"><span v-if="props.row._guardrail_categories && props.row._guardrail_categories.length" style="font-weight:700;color:#616161">Categories: </span><span v-if="props.row._guardrail_categories && props.row._guardrail_categories.length" style="color:#374151">{{ props.row._guardrail_categories.join(', ') }}</span><span v-if="props.row._guardrail_categories && props.row._guardrail_categories.length">&#10;&#10;</span><span style="font-weight:700;color:#616161">Explanation: </span><span style="color:#6b7280">{{ props.row._guardrail_explanation }}</span></pre>
-      </div>
-    </div>
-  </q-td>
-</q-tr>
-""",
-                    )
-
-                    ui.separator().classes("mt-2")
-                    with ui.row().classes("items-center gap-2 mt-2 flex-wrap"):
-                        ui.label("Summary:").classes(
-                            "text-xs font-semibold text-grey-6"
-                        )
-                        ui.label(
-                            f"{len(prefix_rows)} prefix{'es' if len(prefix_rows) != 1 else ''}"
-                        ).classes("text-xs text-grey-6")
-                        if n_jailbreaks:
-                            ui.badge(
-                                f"{n_jailbreaks} Jailbreak{'s' if n_jailbreaks != 1 else ''}",
-                                color="negative",
-                            ).classes("text-xs")
-                        if n_mitigated:
-                            ui.badge(
-                                f"{n_mitigated} Mitigated", color="positive"
-                            ).classes("text-xs")
-                        if n_errors:
-                            ui.badge(
-                                f"{n_errors} Error{'s' if n_errors != 1 else ''}",
-                                color="warning",
-                            ).classes("text-xs")
-                        if n_pending:
-                            ui.badge(f"{n_pending} Pending", color="grey").classes(
-                                "text-xs"
-                            )
-
-                if not detail_mode:
-                    self._wire_expand_toggle(body_col)
-
-    # ── PAP ───────────────────────────────────────────────────────────────────
-
-    @staticmethod
-    def _parse_pap_traces(traces: list[dict]) -> list[dict]:
-        """Parse PAP traces into per-technique rows for the result table."""
-        candidates: dict[int, dict] = {}
-        evaluations: dict[int, dict] = {}
-
-        for td in sorted(traces, key=lambda x: x.get("sequence", 0)):
-            content = td.get("content")
-            if not isinstance(content, dict):
-                continue
-            meta = content.get("metadata") or {}
-            display_type = meta.get("display_type") or ""
-            tech_idx = meta.get("technique_index")
-            if tech_idx is None:
-                continue
-            if display_type == "pap_candidate":
-                req = content.get("request") or {}
-                prompt = req.get("prompt") or "" if isinstance(req, dict) else ""
-                # If prompt is empty, check if there's messages format
-                if not prompt and isinstance(req, dict):
-                    msgs = req.get("messages") or []
-                    for m in reversed(msgs):
-                        if isinstance(m, dict) and m.get("role") == "user":
-                            prompt = str(m.get("content") or "")
-                            break
-                _cand_resp = content.get("response")
-                _cand_resp_actual, _cand_g_side, _cand_g_expl, _cand_g_cats = (
-                    DashboardPage._extract_guardrail_from_response(_cand_resp)
-                )
-                candidates[tech_idx] = {
-                    "technique": meta.get("technique") or "",
-                    "prompt": prompt,
-                    "response": str(_cand_resp_actual) if _cand_resp_actual else "",
-                    "_guardrail_side": _cand_g_side,
-                    "_guardrail_explanation": _cand_g_expl,
-                    "_guardrail_categories": _cand_g_cats,
-                }
-            elif display_type == "pap_evaluation":
-                _pap_raw_resp = content.get("response")
-                _pap_raw_resp, _pap_g_side, _pap_g_expl, _pap_g_cats = (
-                    DashboardPage._extract_guardrail_from_response(_pap_raw_resp)
-                )
-                _pap_response = (
-                    _pap_raw_resp.get("target_response")
-                    if isinstance(_pap_raw_resp, dict)
-                    else None
-                )
-                evaluations[tech_idx] = {
-                    "is_jailbreak": bool(meta.get("is_jailbreak")),
-                    "judge_score": meta.get("judge_score"),
-                    "response": _pap_response or "",
-                    "_guardrail_side": _pap_g_side,
-                    "_guardrail_explanation": _pap_g_expl,
-                }
-
-        rows = []
-        for idx in sorted(candidates):
-            cand = candidates[idx]
-            ev = evaluations.get(idx, {})
-            technique = cand["technique"]
-            prompt = cand["prompt"]
-            is_jailbreak = ev.get("is_jailbreak", False)
-            response = ev.get("response") or cand.get("response") or ""
-            _guardrail_side = (
-                ev.get("_guardrail_side") or cand.get("_guardrail_side") or ""
-            )
-            _guardrail_explanation = (
-                ev.get("_guardrail_explanation")
-                or cand.get("_guardrail_explanation")
-                or ""
-            )
-            if _guardrail_side:
-                bucket = "mitigated"
-            elif is_jailbreak:
-                bucket = "jailbreak"
-            elif ev:
-                bucket = "mitigated"
-            else:
-                bucket = "error"
-            rows.append(
-                {
-                    "num": idx + 1,
-                    "technique": technique,
-                    "prompt_short": (prompt[:80] + "\u2026")
-                    if len(prompt) > 80
-                    else prompt,
-                    "result": "Jailbreak"
-                    if bucket == "jailbreak"
-                    else "Mitigated"
-                    if bucket == "mitigated"
-                    else "Error",
-                    "_bucket": bucket,
-                    "_full_prompt": prompt,
-                    "_response": response,
-                    "_guardrail_side": _guardrail_side,
-                    "_guardrail_explanation": _guardrail_explanation,
-                }
-            )
-        return rows
-
-    def _render_pap_goal_card(
-        self, row: dict, technique_rows: list[dict], detail_mode: bool = False
-    ) -> None:
-        """Render a per-goal PAP result card with a per-technique table."""
-        with self._goal_card_shell(row, detail_mode):
-            if not technique_rows:
-                ui.label("No PAP technique results recorded.").classes(
-                    "text-sm text-grey-6"
-                )
-            else:
-                with ui.column().classes("w-full gap-2 mt-1") as body_col:
-                    if not detail_mode:
-                        body_col.set_visibility(False)
-
-                    pap_cols = [
-                        {
-                            "name": "technique",
-                            "label": "Technique",
-                            "field": "technique",
-                            "align": "left",
-                        },
-                        {
-                            "name": "result",
-                            "label": "Result",
-                            "field": "result",
-                            "align": "center",
-                            "style": "width:100px",
-                        },
-                    ]
-
-                    pap_tbl = (
-                        ui.table(columns=pap_cols, rows=technique_rows, row_key="num")
-                        .classes("w-full text-xs")
-                        .props("dense flat")
-                    )
-                    if detail_mode:
-                        pap_tbl.props(
-                            f":expanded-rows='{json.dumps([r['num'] for r in technique_rows])}'"
-                        )
-
-                    pap_tbl.add_slot(
-                        "body",
-                        r"""
-<q-tr :props="props" @click="props.expand = !props.expand" style="cursor:pointer">
-  <q-td key="technique" :props="props"
-        style="white-space:pre-wrap;word-break:break-word;max-width:360px">
-    {{ props.row.technique }}
-  </q-td>
-  <q-td key="result" :props="props">
-    <q-badge v-if="props.row._bucket === 'jailbreak'" color="negative" class="text-xs">Jailbreak</q-badge>
-    <q-badge v-else-if="props.row._bucket === 'mitigated'" color="positive" class="text-xs">Mitigated</q-badge>
-    <q-badge v-else color="warning" class="text-xs">Error</q-badge>
-  </q-td>
-</q-tr>
-<q-tr v-show="props.expand" :props="props" @click="props.expand = false" style="cursor:pointer">
-  <q-td colspan="100%" class="bg-grey-1">
-    <div class="q-pa-sm">
-      <div class="row items-center justify-between q-mb-xs">
-        <span class="text-caption text-weight-bold text-uppercase text-grey-6">PROMPT SENT TO TARGET</span>
-        <span v-if="props.row._full_prompt" :data-cliptext="props.row._full_prompt||''" onclick="event.stopPropagation();var s=this,t=s.dataset.cliptext||'',ic=s.querySelector('.q-icon');if(window.navigator&&window.navigator.clipboard)window.navigator.clipboard.writeText(t);if(ic)ic.textContent='check';s.setAttribute('title','Copied to clipboard');setTimeout(function(){if(ic)ic.textContent='content_copy';s.setAttribute('title','Copy to clipboard');},2000);" style="display:inline-flex;cursor:pointer" title="Copy to clipboard"><q-btn flat dense size="xs" icon="content_copy" color="grey-6" /></span>
-      </div>
-      <pre v-if="props.row._full_prompt" style="font-size:11px;padding:8px;background:white;border:1px solid #e0e0e0;
-                  border-radius:4px;margin-bottom:8px;white-space:pre-wrap;word-break:break-word">{{ props.row._full_prompt }}</pre>
-      <div v-else class="text-caption text-italic text-grey-5 q-mb-sm">Attacker failed to generate a persuasive prompt for this technique.</div>
-      <template v-if="props.row._full_prompt && props.row._guardrail_side !== 'before'">
-        <div class="row items-center justify-between q-mb-xs">
-          <span class="text-caption text-weight-bold text-uppercase text-grey-6">TARGET RESPONSE</span>
-          <span :data-cliptext="props.row._response||''" onclick="event.stopPropagation();var s=this,t=s.dataset.cliptext||'',ic=s.querySelector('.q-icon');if(window.navigator&&window.navigator.clipboard)window.navigator.clipboard.writeText(t);if(ic)ic.textContent='check';s.setAttribute('title','Copied to clipboard');setTimeout(function(){if(ic)ic.textContent='content_copy';s.setAttribute('title','Copy to clipboard');},2000);" style="display:inline-flex;cursor:pointer" title="Copy to clipboard"><q-btn flat dense size="xs" icon="content_copy" color="grey-6" /></span>
-        </div>
-        <pre style="font-size:11px;padding:8px;background:white;border:1px solid #e0e0e0;
-                    border-radius:4px;margin-bottom:8px;white-space:pre-wrap;word-break:break-word">{{ props.row._response || 'No response recorded.' }}</pre>
-      </template>
-      <div v-if="props.row._guardrail_side === 'before'" style="margin-bottom:8px">
-        <div class="text-caption text-weight-bold text-uppercase q-mb-xs" style="color:#c2410c">⚠ BEFORE GUARDRAIL — BLOCKED</div>
-        <pre style="font-size:11px;padding:10px;background:#fff7ed;border:2px solid #f97316;border-radius:4px;white-space:pre-wrap;word-break:break-word;margin:0"><span v-if="props.row._guardrail_categories && props.row._guardrail_categories.length" style="font-weight:700;color:#c2410c">Categories: </span><span v-if="props.row._guardrail_categories && props.row._guardrail_categories.length" style="color:#9a3412">{{ props.row._guardrail_categories.join(', ') }}</span><span v-if="props.row._guardrail_categories && props.row._guardrail_categories.length">&#10;&#10;</span><span style="font-weight:700;color:#c2410c">Explanation: </span><span style="color:#6b7280">{{ props.row._guardrail_explanation }}</span></pre>
-      </div>
-      <div v-else-if="props.row._guardrail_side === 'after'" style="margin-bottom:8px">
-        <div class="text-caption text-weight-bold text-uppercase q-mb-xs" style="color:#dc2626">🚫 AFTER GUARDRAIL — CENSORED</div>
-        <pre style="font-size:11px;padding:10px;background:#fef2f2;border:2px solid #ef4444;border-radius:4px;white-space:pre-wrap;word-break:break-word;margin:0"><span v-if="props.row._guardrail_categories && props.row._guardrail_categories.length" style="font-weight:700;color:#dc2626">Categories: </span><span v-if="props.row._guardrail_categories && props.row._guardrail_categories.length" style="color:#991b1b">{{ props.row._guardrail_categories.join(', ') }}</span><span v-if="props.row._guardrail_categories && props.row._guardrail_categories.length">&#10;&#10;</span><span style="font-weight:700;color:#dc2626">Explanation: </span><span style="color:#6b7280">{{ props.row._guardrail_explanation }}</span></pre>
-      </div>
-      <div v-else-if="props.row._guardrail_side" style="margin-bottom:8px">
-        <div class="text-caption text-weight-bold text-uppercase q-mb-xs" style="color:#616161">🛡 GUARDRAIL — BLOCKED</div>
-        <pre style="font-size:11px;padding:10px;background:#f5f5f5;border:2px solid #9e9e9e;border-radius:4px;white-space:pre-wrap;word-break:break-word;margin:0"><span v-if="props.row._guardrail_categories && props.row._guardrail_categories.length" style="font-weight:700;color:#616161">Categories: </span><span v-if="props.row._guardrail_categories && props.row._guardrail_categories.length" style="color:#374151">{{ props.row._guardrail_categories.join(', ') }}</span><span v-if="props.row._guardrail_categories && props.row._guardrail_categories.length">&#10;&#10;</span><span style="font-weight:700;color:#616161">Explanation: </span><span style="color:#6b7280">{{ props.row._guardrail_explanation }}</span></pre>
-      </div>
-    </div>
-  </q-td>
-</q-tr>
-""",
-                    )
-
-                if not detail_mode:
-                    self._wire_expand_toggle(body_col)
-
-    # ── TAP ───────────────────────────────────────────────────────────────────
-
-    @staticmethod
-    def _parse_tap_traces(traces: list[dict]) -> tuple[list[dict], dict[int, dict]]:
-        """Parse TAP traces into a list of candidate node dicts."""
-        nodes: list[dict] = []
-        seen_ids: set[str] = set()
-        _interaction_counts: dict[int, int] = {}
-        _summary_counts: dict[int, int] = {}
-
-        def _add(node: dict) -> None:
-            sid = node.get("self_id") or ""
-            if sid and sid in seen_ids:
-                return
-            if sid:
-                seen_ids.add(sid)
-            nodes.append(node)
-
-        for td in sorted(traces, key=lambda x: x.get("sequence", 0)):
-            content = td.get("content")
-            if not isinstance(content, dict):
-                continue
-
-            step_name = str(content.get("step_name") or "")
-
-            if "Depth" in step_name and "Candidate" in step_name:
-                meta = content.get("metadata") or {}
-                req = content.get("request") or {}
-                prompt = req.get("prompt", "") if isinstance(req, dict) else ""
-                resp = content.get("response")
-                resp, _tap_g_side, _tap_g_expl, _tap_g_cats = (
-                    DashboardPage._extract_guardrail_from_response(resp)
-                )
-                response = str(resp) if resp not in (None, "") else ""
-                depth_level = int(meta.get("iteration") or 0)
-                _interaction_counts[depth_level] = (
-                    _interaction_counts.get(depth_level, 0) + 1
-                )
-                _add(
-                    {
-                        "depth": depth_level,
-                        "branch_index": meta.get("branch_index"),
-                        "stream_index": meta.get("stream_index"),
-                        "self_id": meta.get("self_id", ""),
-                        "parent_id": meta.get("parent_id"),
-                        "prompt": prompt,
-                        "response": response,
-                        "judge_score": meta.get("judge_score"),
-                        "on_topic": meta.get("on_topic_score"),
-                        "improvement": str(meta.get("improvement") or ""),
-                        "_guardrail_side": _tap_g_side,
-                        "_guardrail_explanation": _tap_g_expl,
-                        "_guardrail_categories": _tap_g_cats,
-                    }
-                )
-                continue
-
-            if "Depth" in step_name and "Summary" in step_name:
-                depth_level = int(content.get("depth") or 0)
-                branches = [
-                    b for b in (content.get("branches") or []) if isinstance(b, dict)
-                ]
-                _summary_counts[depth_level] = len(branches)
-                for branch in branches:
-                    _b_resp = branch.get("response")
-                    _b_resp, _b_g_side, _b_g_expl, _b_g_cats = (
-                        DashboardPage._extract_guardrail_from_response(_b_resp)
-                    )
-                    _add(
-                        {
-                            "depth": depth_level,
-                            "branch_index": branch.get("branch_index"),
-                            "stream_index": branch.get("stream_index"),
-                            "self_id": branch.get("self_id", ""),
-                            "parent_id": branch.get("parent_id"),
-                            "prompt": str(branch.get("prompt") or ""),
-                            "response": str(_b_resp or ""),
-                            "judge_score": branch.get("judge_score"),
-                            "on_topic": branch.get("on_topic_score"),
-                            "improvement": str(branch.get("improvement") or ""),
-                            "_guardrail_side": _b_g_side,
-                            "_guardrail_explanation": _b_g_expl,
-                            "_guardrail_categories": _b_g_cats,
-                        }
-                    )
-                continue
-
-            if not step_name and "depth" in content and "branches" in content:
-                depth_level = int(content.get("depth") or 0)
-                branches = [
-                    b for b in (content.get("branches") or []) if isinstance(b, dict)
-                ]
-                _summary_counts[depth_level] = len(branches)
-                for branch in branches:
-                    _b_resp2 = branch.get("response")
-                    _b_resp2, _b2_g_side, _b2_g_expl, _b2_g_cats = (
-                        DashboardPage._extract_guardrail_from_response(_b_resp2)
-                    )
-                    _add(
-                        {
-                            "depth": depth_level,
-                            "branch_index": branch.get("branch_index"),
-                            "stream_index": branch.get("stream_index"),
-                            "self_id": branch.get("self_id", ""),
-                            "parent_id": branch.get("parent_id"),
-                            "prompt": str(branch.get("prompt") or ""),
-                            "response": str(_b_resp2 or ""),
-                            "judge_score": branch.get("judge_score"),
-                            "on_topic": branch.get("on_topic_score"),
-                            "improvement": str(branch.get("improvement") or ""),
-                            "_guardrail_side": _b2_g_side,
-                            "_guardrail_explanation": _b2_g_expl,
-                            "_guardrail_categories": _b2_g_cats,
-                        }
-                    )
-                continue
-
-        if not nodes:
-            eval_idx = 0
-            for td in sorted(traces, key=lambda x: x.get("sequence", 0)):
-                content = td.get("content")
-                if not isinstance(content, dict):
-                    continue
-                if content.get("step_name") != "Evaluation":
-                    continue
-                evaluator = str(content.get("evaluator") or "")
-                if evaluator == "tracking_coordinator":
-                    continue
-                meta = content.get("metadata") or {}
-                prompt = str(meta.get("prefix") or "")
-                response = str(meta.get("completion") or "")
-                if not prompt and not response:
-                    continue
-                score_raw = content.get("score")
-                try:
-                    score_val = int(float(score_raw)) if score_raw is not None else None
-                except (TypeError, ValueError):
-                    score_val = None
-                # Binary evaluators (HarmBench, JailbreakBench) return 0/1;
-                # normalize to the 1–10 TAP scale so display is consistent
-                # with scores from Candidate traces.
-                _ev_lower = evaluator.lower()
-                if score_val is not None and (
-                    "harmbench" in _ev_lower or "jailbreakbench" in _ev_lower
-                ):
-                    score_val = 1 if score_val == 0 else 10
-                eval_idx += 1
-                nodes.append(
-                    {
-                        "depth": 0,
-                        "branch_index": eval_idx - 1,
-                        "stream_index": 0,
-                        "self_id": "",
-                        "parent_id": None,
-                        "prompt": prompt,
-                        "response": response,
-                        "judge_score": score_val,
-                        "on_topic": None,
-                        "improvement": "",
-                        "_guardrail_side": "",
-                        "_guardrail_explanation": "",
-                        "_guardrail_categories": [],
-                    }
-                )
-
-        depth_stats: dict[int, dict] = {}
-        all_depths = set(_interaction_counts) | set(_summary_counts)
-        for _d in all_depths:
-            _gen = _interaction_counts.get(_d)
-            _surv = _summary_counts.get(_d)
-            depth_stats[_d] = {
-                "generated": _gen,
-                "survived": _surv,
-                "pruned": (_gen - _surv)
-                if (_gen is not None and _surv is not None)
-                else None,
-            }
-
-        return nodes, depth_stats
-
-    def _render_tap_goal_card(
-        self,
-        row: dict,
-        nodes: list[dict],
-        depth_stats: dict[int, dict] | None = None,
-        detail_mode: bool = False,
-    ) -> None:
-        """Render a TAP goal card: one table per depth."""
-        with self._goal_card_shell(row, detail_mode):
-            if not nodes:
-                ui.label("No TAP candidate data recorded.").classes(
-                    "text-sm text-grey-6"
-                )
-            else:
-                by_depth: dict[int, list[dict]] = defaultdict(list)
-                for n in nodes:
-                    by_depth[n.get("depth") or 0].append(n)
-
-                _global_num = 0
-                _id_to_num: dict[str, int] = {}
-                for depth_level in sorted(by_depth.keys()):
-                    depth_nodes = by_depth[depth_level]
-                    depth_nodes.sort(
-                        key=lambda x: (
-                            x.get("stream_index") or 0,
-                            x.get("branch_index") or 0,
-                        )
-                    )
-                    for n in depth_nodes:
-                        _global_num += 1
-                        n["_global_num"] = _global_num
-                        sid = n.get("self_id") or ""
-                        if sid:
-                            _id_to_num[sid] = _global_num
-                for n in nodes:
-                    pid = n.get("parent_id") or ""
-                    n["_parent_num"] = _id_to_num.get(pid)
-
-                with ui.column().classes("w-full gap-2 mt-1") as body_col:
-                    if not detail_mode:
-                        body_col.set_visibility(False)
-
-                    for depth_level in sorted(by_depth.keys()):
-                        depth_nodes = by_depth[depth_level]
-                        _ds = (depth_stats or {}).get(depth_level, {})
-                        _n_cands = len(depth_nodes)
-                        _cand_label = f"{_n_cands} candidate{'s' if _n_cands != 1 else ''} after pruning"
-                        _depth_header = (
-                            "Final Evaluation"
-                            if depth_level == 0
-                            else f"Depth {depth_level}"
-                        )
-                        with ui.row().classes("items-center gap-2 mt-3 mb-0.5 px-1"):
-                            ui.label(_depth_header).classes(
-                                "text-xs font-semibold text-grey-6 uppercase tracking-wide"
-                            )
-                            ui.badge(_cand_label, color="grey-5").classes("text-xs")
-
-                        rows_data = []
-                        for idx, n in enumerate(depth_nodes):
-                            score = n.get("judge_score")
-                            on_topic = n.get("on_topic")
-                            prompt_text = n.get("prompt") or ""
-                            parent_num = n.get("_parent_num")
-                            _s_str = (
-                                f"{int(float(score))}/10" if score is not None else "—"
-                            )
-                            rows_data.append(
-                                {
-                                    "_num": n["_global_num"],
-                                    "parent_str": str(parent_num)
-                                    if parent_num is not None
-                                    else "—",
-                                    "prompt_short": (prompt_text[:80] + "…")
-                                    if len(prompt_text) > 80
-                                    else prompt_text or "—",
-                                    "score_val": float(score)
-                                    if score is not None
-                                    else -1,
-                                    "score_str": _s_str,
-                                    "on_topic_str": (
-                                        "yes"
-                                        if on_topic is not None
-                                        and int(float(on_topic)) >= 1
-                                        else "no"
-                                        if on_topic is not None
-                                        else "—"
-                                    ),
-                                    "_on_topic": on_topic,
-                                    "_full_prompt": prompt_text,
-                                    "_response": n.get("response") or "",
-                                    "_guardrail_side": n.get("_guardrail_side") or "",
-                                    "_guardrail_explanation": n.get(
-                                        "_guardrail_explanation"
-                                    )
-                                    or "",
-                                }
-                            )
-
-                        columns = [
-                            {
-                                "name": "_num",
-                                "label": "#",
-                                "field": "_num",
-                                "align": "center",
-                                "style": "width:40px",
-                            },
-                            {
-                                "name": "parent_str",
-                                "label": "Parent",
-                                "field": "parent_str",
-                                "align": "center",
-                                "style": "width:55px",
-                            },
-                            {
-                                "name": "prompt_short",
-                                "label": "Prompt",
-                                "field": "prompt_short",
-                                "align": "left",
-                            },
-                            {
-                                "name": "on_topic_str",
-                                "label": "On-topic",
-                                "field": "on_topic_str",
-                                "align": "center",
-                                "style": "width:80px",
-                            },
-                            {
-                                "name": "score_str",
-                                "label": "Score",
-                                "field": "score_str",
-                                "align": "center",
-                                "style": "width:80px",
-                            },
-                        ]
-
-                        tbl = (
-                            ui.table(columns=columns, rows=rows_data, row_key="_num")
-                            .classes("w-full text-xs")
-                            .props("dense flat")
-                        )
-                        if detail_mode:
-                            tbl.props(
-                                f":expanded-rows='{json.dumps([r['_num'] for r in rows_data])}'"
-                            )
-
-                        tbl.add_slot(
-                            "body",
-                            r"""
-<q-tr :props="props" @click="props.expand = !props.expand" style="cursor:pointer">
-  <q-td key="_num" :props="props" class="text-center text-grey-6">{{ props.row._num }}</q-td>
-  <q-td key="parent_str" :props="props" class="text-center text-grey-5">{{ props.row.parent_str }}</q-td>
-  <q-td key="prompt_short" :props="props"
-        style="white-space:pre-wrap;word-break:break-word;max-width:360px">
-    {{ props.row.prompt_short }}
-  </q-td>
-  <q-td key="on_topic_str" :props="props" class="text-center">
-    <span class="text-grey-7">{{ props.row.on_topic_str }}</span>
-  </q-td>
-  <q-td key="score_str" :props="props" class="text-center">
-    <span class="text-grey-7">{{ props.row.score_str }}</span>
-  </q-td>
-</q-tr>
-<q-tr v-show="props.expand" :props="props" @click="props.expand = false" style="cursor:pointer">
-  <q-td colspan="100%" class="bg-grey-1">
-    <div class="q-pa-sm">
-      <div class="row items-center justify-between q-mb-xs">
-        <span class="text-caption text-weight-bold text-uppercase text-grey-6">PROMPT SENT TO TARGET</span>
-        <span :data-cliptext="props.row._full_prompt||''" onclick="event.stopPropagation();var s=this,t=s.dataset.cliptext||'',ic=s.querySelector('.q-icon');if(window.navigator&&window.navigator.clipboard)window.navigator.clipboard.writeText(t);if(ic)ic.textContent='check';s.setAttribute('title','Copied to clipboard');setTimeout(function(){if(ic)ic.textContent='content_copy';s.setAttribute('title','Copy to clipboard');},2000);" style="display:inline-flex;cursor:pointer" title="Copy to clipboard"><q-btn flat dense size="xs" icon="content_copy" color="grey-6" /></span>
-      </div>
-      <pre style="font-size:11px;padding:8px;background:white;border:1px solid #e0e0e0;
-                  border-radius:4px;margin-bottom:8px;white-space:pre-wrap;word-break:break-word">{{ props.row._full_prompt || '\u2014' }}</pre>
-      <template v-if="props.row._guardrail_side !== 'before'">
-        <div class="row items-center justify-between q-mb-xs">
-          <span class="text-caption text-weight-bold text-uppercase text-grey-6">TARGET RESPONSE</span>
-          <span :data-cliptext="props.row._response||''" onclick="event.stopPropagation();var s=this,t=s.dataset.cliptext||'',ic=s.querySelector('.q-icon');if(window.navigator&&window.navigator.clipboard)window.navigator.clipboard.writeText(t);if(ic)ic.textContent='check';s.setAttribute('title','Copied to clipboard');setTimeout(function(){if(ic)ic.textContent='content_copy';s.setAttribute('title','Copy to clipboard');},2000);" style="display:inline-flex;cursor:pointer" title="Copy to clipboard"><q-btn flat dense size="xs" icon="content_copy" color="grey-6" /></span>
-        </div>
-        <pre style="font-size:11px;padding:8px;background:white;border:1px solid #e0e0e0;
-                    border-radius:4px;margin-bottom:8px;white-space:pre-wrap;word-break:break-word">{{ props.row._response || 'No response recorded.' }}</pre>
-      </template>
-      <div v-if="props.row._guardrail_side === 'before'" style="margin-bottom:8px">
-        <div class="text-caption text-weight-bold text-uppercase q-mb-xs" style="color:#c2410c">⚠ BEFORE GUARDRAIL — BLOCKED</div>
-        <pre style="font-size:11px;padding:10px;background:#fff7ed;border:2px solid #f97316;border-radius:4px;white-space:pre-wrap;word-break:break-word;margin:0"><span v-if="props.row._guardrail_categories && props.row._guardrail_categories.length" style="font-weight:700;color:#c2410c">Categories: </span><span v-if="props.row._guardrail_categories && props.row._guardrail_categories.length" style="color:#9a3412">{{ props.row._guardrail_categories.join(', ') }}</span><span v-if="props.row._guardrail_categories && props.row._guardrail_categories.length">&#10;&#10;</span><span style="font-weight:700;color:#c2410c">Explanation: </span><span style="color:#6b7280">{{ props.row._guardrail_explanation }}</span></pre>
-      </div>
-      <div v-else-if="props.row._guardrail_side === 'after'" style="margin-bottom:8px">
-        <div class="text-caption text-weight-bold text-uppercase q-mb-xs" style="color:#dc2626">🚫 AFTER GUARDRAIL — CENSORED</div>
-        <pre style="font-size:11px;padding:10px;background:#fef2f2;border:2px solid #ef4444;border-radius:4px;white-space:pre-wrap;word-break:break-word;margin:0"><span v-if="props.row._guardrail_categories && props.row._guardrail_categories.length" style="font-weight:700;color:#dc2626">Categories: </span><span v-if="props.row._guardrail_categories && props.row._guardrail_categories.length" style="color:#991b1b">{{ props.row._guardrail_categories.join(', ') }}</span><span v-if="props.row._guardrail_categories && props.row._guardrail_categories.length">&#10;&#10;</span><span style="font-weight:700;color:#dc2626">Explanation: </span><span style="color:#6b7280">{{ props.row._guardrail_explanation }}</span></pre>
-      </div>
-      <div v-else-if="props.row._guardrail_side" style="margin-bottom:8px">
-        <div class="text-caption text-weight-bold text-uppercase q-mb-xs" style="color:#616161">🛡 GUARDRAIL — BLOCKED</div>
-        <pre style="font-size:11px;padding:10px;background:#f5f5f5;border:2px solid #9e9e9e;border-radius:4px;white-space:pre-wrap;word-break:break-word;margin:0"><span v-if="props.row._guardrail_categories && props.row._guardrail_categories.length" style="font-weight:700;color:#616161">Categories: </span><span v-if="props.row._guardrail_categories && props.row._guardrail_categories.length" style="color:#374151">{{ props.row._guardrail_categories.join(', ') }}</span><span v-if="props.row._guardrail_categories && props.row._guardrail_categories.length">&#10;&#10;</span><span style="font-weight:700;color:#616161">Explanation: </span><span style="color:#6b7280">{{ props.row._guardrail_explanation }}</span></pre>
-      </div>
-    </div>
-  </q-td>
-</q-tr>
-""",
-                        )
-
-                if not detail_mode:
-                    self._wire_expand_toggle(body_col)
-
-    # ── Generic / fallback ────────────────────────────────────────────────────
-
-    @staticmethod
-    def _extract_prompt_response_from_traces(
-        traces: list[dict],
-    ) -> tuple[str, str, "dict | None"]:
-        """Extract the best (prompt, response, guardrail_event) from generic attack traces."""
-        best_req = ""
-        best_resp = ""
-        fallback_req = ""
-        guardrail_event: dict | None = None
-        for td in sorted(traces, key=lambda x: x.get("sequence", 0)):
-            content = td.get("content")
-            if not isinstance(content, dict):
-                continue
-            req, resp = DashboardPage._extract_request_response_candidates(content)
-            # Detect guardrail event dict before stringifying
-            if isinstance(resp, dict) and resp.get("side") in (
-                "before",
-                "after",
-                "unknown",
-            ):
-                guardrail_event = resp
-                resp = None  # Don't treat guardrail dict as normal response text
-            req_str = (
-                json.dumps(req, indent=2)
-                if isinstance(req, (dict, list))
-                else str(req)
-                if req not in (None, "")
-                else ""
-            )
-            resp_str = (
-                json.dumps(resp, indent=2)
-                if isinstance(resp, (dict, list))
-                else str(resp)
-                if resp not in (None, "")
-                else ""
-            )
-            if req_str and resp_str:
-                best_req = req_str
-                best_resp = resp_str
-            elif req_str:
-                fallback_req = req_str
-        if best_req:
-            return best_req, best_resp, guardrail_event
-        if fallback_req:
-            return fallback_req, "(no response recorded)", guardrail_event
-        return "(not available)", "(not available)", guardrail_event
-
-    def _render_generic_goal_card(
-        self,
-        row: dict,
-        request_text: str,
-        response_text: str,
-        detail_mode: bool = False,
-        guardrail_event: dict | None = None,
-    ) -> None:
-        """Render a per-goal result card for non-specific attacks."""
-        with self._goal_card_shell(row, detail_mode):
-            with ui.column().classes("w-full gap-2 mt-1") as body_col:
-                if not detail_mode:
-                    body_col.set_visibility(False)
-
-                with ui.row().classes("w-full items-center justify-between"):
-                    ui.label("PROMPT SENT TO TARGET").classes(
-                        "text-[10px] text-grey-6 font-semibold uppercase tracking-wide"
-                    )
-                    ui.button(icon="content_copy").props(
-                        "flat dense size=xs color=grey-6"
-                    ).tooltip("Copy to clipboard").on(
-                        "click",
-                        js_handler=f"(event) => {{var b=event.currentTarget,ic=b.querySelector('.q-icon');if(navigator.clipboard)navigator.clipboard.writeText({json.dumps(request_text or '')});if(ic){{ic.textContent='check';setTimeout(function(){{ic.textContent='content_copy';}},2000);}}}}",
-                    )
-                ui.html(
-                    '<pre style="font-size:11px;padding:8px;background:white;'
-                    "border:1px solid #e0e0e0;border-radius:4px;margin-bottom:8px;"
-                    'white-space:pre-wrap;word-break:break-word">'
-                    + html.escape(request_text or "\u2014")
-                    + "</pre>"
-                )
-
-                _g_side = (guardrail_event or {}).get("side") or ""
-                if _g_side == "before":
-                    self._render_guardrail_event_block(guardrail_event)  # type: ignore[arg-type]
-                else:
-                    with ui.row().classes("w-full items-center justify-between"):
-                        ui.label("TARGET RESPONSE").classes(
-                            "text-[10px] text-grey-6 font-semibold uppercase tracking-wide"
-                        )
-                        ui.button(icon="content_copy").props(
-                            "flat dense size=xs color=grey-6"
-                        ).tooltip("Copy to clipboard").on(
-                            "click",
-                            js_handler=f"(event) => {{var b=event.currentTarget,ic=b.querySelector('.q-icon');if(navigator.clipboard)navigator.clipboard.writeText({json.dumps(response_text or '')});if(ic){{ic.textContent='check';setTimeout(function(){{ic.textContent='content_copy';}},2000);}}}}",
-                        )
-                    ui.html(
-                        '<pre style="font-size:11px;padding:8px;background:white;'
-                        "border:1px solid #e0e0e0;border-radius:4px;"
-                        'white-space:pre-wrap;word-break:break-word">'
-                        + html.escape(response_text or "No response recorded.")
-                        + "</pre>"
-                    )
-                    if _g_side:
-                        self._render_guardrail_event_block(guardrail_event)  # type: ignore[arg-type]
-
-            if not detail_mode:
-                self._wire_expand_toggle(body_col)
-
     async def _load_attack_specific_traces(
         self, row: dict, container: ui.column, attack_str: str
     ) -> None:
@@ -6557,6 +3970,9 @@ function hackAgentCopyFallback(text) {
                 elif atk == "autodanturbo":
                     detail_data = self._parse_autodan_traces(serialized_traces)
                     self._render_autodan_goal_card(row, detail_data, detail_mode=True)
+                elif atk == "mml":
+                    detail_data = self._parse_mml_traces(serialized_traces)
+                    self._render_mml_goal_card(row, detail_data, detail_mode=True)
                 else:
                     req_text, resp_text, _generic_guardrail = (
                         self._extract_prompt_response_from_traces(serialized_traces)
@@ -7992,6 +5408,9 @@ function hackAgentCopyFallback(text) {
                     color = "positive" if bool(success) else "warning"
                     ui.badge(label, color=color).classes("text-xs")
 
+            # MML: render encoded image inline if present in metadata
+            self._render_mml_trace_image(metadata)
+
             for title, value in blocks:
                 if value is None or value == "":
                     continue
@@ -8077,72 +5496,6 @@ function hackAgentCopyFallback(text) {
             return
 
         ui.label(str(content)).classes("text-sm whitespace-pre-wrap")
-
-    @staticmethod
-    def _render_guardrail_event_block(event: dict) -> None:
-        """Render a visual banner for a guardrail-blocked trace step.
-
-        * ``side="before"`` — prompt was blocked before reaching the target model:
-          shows an orange warning box.  No target response is displayed because
-          the request was never sent.
-        * ``side="after"`` — model response was censored after it was received:
-          shows a red box below the (visible) target response.
-        """
-        side = event.get("side", "unknown")
-        explanation = str(
-            event.get("explanation")
-            or event.get("reasoning")
-            or event.get("message")
-            or "Blocked by guardrail"
-        )
-        categories = event.get("categories", [])
-
-        cat_html = ""
-        if categories:
-            cat_str = ", ".join(str(c) for c in categories)
-            if side == "before":
-                cat_html = f'<span style="font-weight:700;color:#c2410c">Categories: </span><span style="color:#9a3412">{cat_str}</span><br><br>'
-            elif side == "after":
-                cat_html = f'<span style="font-weight:700;color:#dc2626">Categories: </span><span style="color:#991b1b">{cat_str}</span><br><br>'
-            else:
-                cat_html = f'<span style="font-weight:700;color:#616161">Categories: </span><span style="color:#374151">{cat_str}</span><br><br>'
-
-        if side == "before":
-            heading = "⚠ BEFORE GUARDRAIL — BLOCKED"
-            border_color = "#f97316"
-            bg_color = "#fff7ed"
-            heading_color = "#c2410c"
-            expl_label = (
-                '<span style="font-weight:700;color:#c2410c">Explanation: </span>'
-            )
-        elif side == "after":
-            heading = "🚫 AFTER GUARDRAIL — CENSORED"
-            border_color = "#ef4444"
-            bg_color = "#fef2f2"
-            heading_color = "#dc2626"
-            expl_label = (
-                '<span style="font-weight:700;color:#dc2626">Explanation: </span>'
-            )
-        else:
-            heading = "🛡 GUARDRAIL — BLOCKED"
-            border_color = "#9e9e9e"
-            bg_color = "#f5f5f5"
-            heading_color = "#616161"
-            expl_label = (
-                '<span style="font-weight:700;color:#616161">Explanation: </span>'
-            )
-
-        from html import escape
-
-        ui.html(
-            f'<div style="margin-bottom:8px">'
-            f'<div style="font-size:11px;font-weight:700;text-transform:uppercase;margin-bottom:4px;color:{heading_color}">{heading}</div>'
-            f'<pre style="font-size:11px;padding:10px;background:{bg_color};border:2px solid {border_color};border-radius:4px;white-space:pre-wrap;word-break:break-word;margin:0">'
-            f"{cat_html}"
-            f"{expl_label}"
-            f'<span style="color:#6b7280">{escape(explanation)}</span>'
-            f"</pre></div>"
-        )
 
     async def refresh_view(self) -> None:
         _v = self.current_view["value"]
@@ -10534,6 +7887,7 @@ function hackAgentCopyFallback(text) {
                     "flipattack": "FlipAttack",
                     "pap": "PAP",
                     "h4rm3l": "H4rm3l",
+                    "mml": "MML",
                 }
 
                 def _h_resolve_eval_label(
@@ -10660,6 +8014,12 @@ function hackAgentCopyFallback(text) {
                                 "Branching",
                                 str(_tap_p.get("branching_factor", 3)),
                             )
+                    elif _atk_lower == "mml":
+                        _mml_params = _hcfg.get("mml_params") or {}
+                        if isinstance(_mml_params, dict):
+                            _mml_enc = _mml_params.get("encoding_mode", "")
+                            if _mml_enc:
+                                _chip("image", "Encoding", str(_mml_enc))
 
                 # ── Line 2: Dataset ───────────────────────────────────
                 _h_dataset_raw = _hcfg.get("dataset") or _hrc.get("dataset")
@@ -11420,6 +8780,9 @@ function hackAgentCopyFallback(text) {
                     elif _h_atk == "autodanturbo":
                         _t = generic_traces_map_hr.get(_rid, [])
                         _h_detail_data[_rid] = self._parse_autodan_traces(_t)
+                    elif _h_atk == "mml":
+                        _t = generic_traces_map_hr.get(_rid, [])
+                        _h_detail_data[_rid] = self._parse_mml_traces(_t)
                     else:
                         _t = generic_traces_map_hr.get(_rid, [])
                         _h_detail_data[_rid] = (
